% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{access_control_req_group}
\alias{access_control_req_group}
\title{Access Control Request for Group}
\usage{
access_control_req_group(
  group,
  permission_level = c("CAN_MANAGE", "CAN_MANAGE_RUN", "CAN_VIEW")
)
}
\arguments{
\item{group}{Group name. There are two built-in groups: \code{users} for all users,
and \code{admins} for administrators.}

\item{permission_level}{Permission level to grant. One of \code{CAN_MANAGE},
\code{CAN_MANAGE_RUN}, \code{CAN_VIEW}.}
}
\description{
Access Control Request for Group
}
\seealso{
\code{\link[=access_control_request]{access_control_request()}}

Other Access Control Request Objects: 
\code{\link{access_control_req_user}()}
}
\concept{Access Control Request Objects}
