% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{access_control_req_user}
\alias{access_control_req_user}
\title{Access Control Request For User}
\usage{
access_control_req_user(
  user_name,
  permission_level = c("CAN_MANAGE", "CAN_MANAGE_RUN", "CAN_VIEW", "IS_OWNER")
)
}
\arguments{
\item{user_name}{Email address for the user.}

\item{permission_level}{Permission level to grant. One of \code{CAN_MANAGE},
\code{CAN_MANAGE_RUN}, \code{CAN_VIEW}, \code{IS_OWNER}.}
}
\description{
Access Control Request For User
}
\seealso{
\code{\link[=access_control_request]{access_control_request()}}

Other Access Control Request Objects: 
\code{\link{access_control_req_group}()}
}
\concept{Access Control Request Objects}
