% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{access_control_request}
\alias{access_control_request}
\title{Access Control Request}
\usage{
access_control_request(...)
}
\arguments{
\item{...}{Instances of \code{\link[=access_control_req_user]{access_control_req_user()}} or
\code{\link[=access_control_req_group]{access_control_req_group()}}.}
}
\description{
Access Control Request
}
\seealso{
\code{\link[=db_cluster_create]{db_cluster_create()}}, \code{\link[=db_cluster_edit]{db_cluster_edit()}}
}
