% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-management.R
\name{add_lib_path}
\alias{add_lib_path}
\title{Add Library Path}
\usage{
add_lib_path(path, after, version = FALSE)
}
\arguments{
\item{path}{Directory that will added as location for which packages
are searched. Recursively creates the directory if it doesn't exist. On
Databricks remember to use \verb{/dbfs/} or \verb{/Volumes/...} as a prefix.}

\item{after}{Location at which to append the \code{path} value after.}

\item{version}{If \code{TRUE} will add the R version string to the end
of \code{path}. This is recommended if using different R versions and sharing a
common \code{path} between users.}
}
\description{
Add Library Path
}
\details{
This functions primary use is when using Databricks notebooks or hosted
RStudio, however, it works anywhere.
}
\seealso{
\code{\link[base:libPaths]{base::.libPaths()}}, \code{\link[=remove_lib_path]{remove_lib_path()}}
}
