% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_browser.R
\name{browse_path}
\alias{browse_path}
\title{Browse a Unity Catalog path}
\usage{
browse_path(path = "/Volumes/", host = db_host(), token = db_token())
}
\arguments{
\item{path}{The initial path to browse}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\value{
An invisible \code{NULL}.
}
\description{
Opens a browser in the RStudio viewer
}
