% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{cluster_autoscale}
\alias{cluster_autoscale}
\title{Cluster Autoscale}
\usage{
cluster_autoscale(min_workers, max_workers)
}
\arguments{
\item{min_workers}{The minimum number of workers to which the cluster can
scale down when underutilized. It is also the initial number of workers the
cluster will have after creation.}

\item{max_workers}{The maximum number of workers to which the cluster can
scale up when overloaded. \code{max_workers} must be strictly greater than
\code{min_workers}.}
}
\description{
Range defining the min and max number of cluster workers.
}
\seealso{
\code{\link[=db_cluster_create]{db_cluster_create()}}, \code{\link[=db_cluster_edit]{db_cluster_edit()}}
}
\concept{Cluster Objects}
