% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{cluster_log_conf}
\alias{cluster_log_conf}
\title{Cluster Log Configuration}
\usage{
cluster_log_conf(dbfs = NULL, s3 = NULL)
}
\arguments{
\item{dbfs}{Instance of \code{\link[=dbfs_storage_info]{dbfs_storage_info()}}.}

\item{s3}{Instance of \code{\link[=s3_storage_info]{s3_storage_info()}}.}
}
\description{
Path to cluster log.
}
\details{
\code{dbfs} and \code{s3} are mutually exclusive, logs can only be sent to
one destination.
}
\seealso{
Other Cluster Log Configuration Objects: 
\code{\link{dbfs_storage_info}()},
\code{\link{s3_storage_info}()}
}
\concept{Cluster Log Configuration Objects}
