% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_list.R
\name{clusters_list}
\alias{clusters_list}
\title{List all clusters in a Databricks workspace.}
\usage{
clusters_list(workspace, token = NULL, verbose = T)
}
\arguments{
\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used.}

\item{verbose}{If true, will pretty print the success or failure of the
request.  Defaults to TRUE.}
}
\value{
A list with two elements:
\itemize{
    \item \emph{response} - The full API response, including status code.
    \item \emph{response_df} - The JSON result transformed into a dataframe.
}
}
\description{
Fetches all of the details for clusters in the authenticated workspace.
}
\details{
The API endpoint for running a job is '2.0/clusters/list'.  For all
details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
clusters <- clusters_list(workspace = "https://eastus2.azuredatabricks.net",
                  token = "dapiafoi32984320890d9asaf1")

## Get active clusters
df <- clusters$df
active <- df[df$state == "RUNNING", c("cluster_id", "state")]
}
