% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{DatabricksSQL}
\alias{DatabricksSQL}
\title{Create Databricks SQL Driver}
\usage{
DatabricksSQL()
}
\value{
A DatabricksDriver object
}
\description{
Create Databricks SQL Driver
}
\examples{
\dontrun{
drv <- DatabricksSQL()
con <- dbConnect(drv, warehouse_id = "your_warehouse_id")
}
}
