% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_append_with_select_values}
\alias{db_append_with_select_values}
\title{Append data using atomic INSERT INTO with SELECT VALUES}
\usage{
db_append_with_select_values(conn, quoted_name, value, insert_sql = NULL)
}
\description{
Append data using atomic INSERT INTO with SELECT VALUES
}
\keyword{internal}
