% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_assert_show_progress}
\alias{db_assert_show_progress}
\title{Assert that a progress flag is valid}
\usage{
db_assert_show_progress(show_progress)
}
\description{
Assert that a progress flag is valid
}
\keyword{internal}
