% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_assert_statement}
\alias{db_assert_statement}
\title{Assert that a statement is provided}
\usage{
db_assert_statement(statement)
}
\description{
Assert that a statement is provided
}
\keyword{internal}
