% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_assert_valid_conn}
\alias{db_assert_valid_conn}
\title{Assert that a connection is valid}
\usage{
db_assert_valid_conn(conn)
}
\description{
Assert that a connection is valid
}
\keyword{internal}
