% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-context.R
\name{db_context_command_parse}
\alias{db_context_command_parse}
\title{Parse Command Results}
\usage{
db_context_command_parse(x, language = c("r", "py", "scala", "sql"))
}
\arguments{
\item{x}{command output from \code{db_context_command_status} or
\code{db_context_manager}'s \code{cmd_run}}

\item{language}{}
}
\value{
command results
}
\description{
Parse Command Results
}
\seealso{
Other Execution Context API: 
\code{\link{db_context_command_cancel}()},
\code{\link{db_context_command_run}()},
\code{\link{db_context_command_run_and_wait}()},
\code{\link{db_context_command_status}()},
\code{\link{db_context_create}()},
\code{\link{db_context_destroy}()},
\code{\link{db_context_status}()}
}
\concept{Execution Context API}
\keyword{internal}
