% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-context.R
\name{db_context_command_run_and_wait}
\alias{db_context_command_run_and_wait}
\title{Run a Command and Wait For Results}
\usage{
db_context_command_run_and_wait(
  cluster_id,
  context_id,
  language = c("python", "sql", "scala", "r"),
  command = NULL,
  command_file = NULL,
  options = list(),
  parse_result = TRUE,
  host = db_host(),
  token = db_token()
)
}
\arguments{
\item{cluster_id}{The ID of the cluster to create the context for.}

\item{context_id}{The ID of the execution context.}

\item{language}{The language for the context. One of \code{python}, \code{sql}, \code{scala},
\code{r}.}

\item{command}{The command string to run.}

\item{command_file}{The path to a file containing the command to run.}

\item{options}{Named list of values used downstream. For example, a
'displayRowLimit' override (used in testing).}

\item{parse_result}{Boolean, determines if results are parsed automatically.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}
}
\value{
Endpoint-specific output.
}
\description{
Run a Command and Wait For Results
}
\seealso{
Other Execution Context API: 
\code{\link{db_context_command_cancel}()},
\code{\link{db_context_command_parse}()},
\code{\link{db_context_command_run}()},
\code{\link{db_context_command_status}()},
\code{\link{db_context_create}()},
\code{\link{db_context_destroy}()},
\code{\link{db_context_status}()}
}
\concept{Execution Context API}
