% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-context.R
\name{db_context_destroy}
\alias{db_context_destroy}
\title{Delete an Execution Context}
\usage{
db_context_destroy(
  cluster_id,
  context_id,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{cluster_id}{The ID of the cluster to create the context for.}

\item{context_id}{The ID of the execution context.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
If \code{perform_request = TRUE}, returns endpoint-specific API output. If \code{FALSE}, returns an \code{httr2_request}.
}
\description{
Delete an Execution Context
}
\seealso{
Other Execution Context API: 
\code{\link{db_context_command_cancel}()},
\code{\link{db_context_command_parse}()},
\code{\link{db_context_command_run}()},
\code{\link{db_context_command_run_and_wait}()},
\code{\link{db_context_command_status}()},
\code{\link{db_context_create}()},
\code{\link{db_context_status}()}
}
\concept{Execution Context API}
