% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_create_table_as_select_values}
\alias{db_create_table_as_select_values}
\title{Create a table or atomically replace it with typed values}
\usage{
db_create_table_as_select_values(
  conn,
  quoted_name,
  value,
  field.types,
  temporary = FALSE,
  overwrite = FALSE
)
}
\description{
Create a table or atomically replace it with typed values
}
\keyword{internal}
