% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_create_table_from_data}
\alias{db_create_table_from_data}
\title{Create table from data frame structure}
\usage{
db_create_table_from_data(
  conn,
  quoted_name,
  value,
  field.types,
  temporary = FALSE,
  overwrite = FALSE
)
}
\description{
Create table from data frame structure
}
\keyword{internal}
