% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{db_current_user}
\alias{db_current_user}
\title{Get Current User Info}
\usage{
db_current_user(host = db_host(), token = db_token(), perform_request = TRUE)
}
\arguments{
\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
list of user metadata
}
\description{
Get Current User Info
}
