% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_escape_string_literal}
\alias{db_escape_string_literal}
\title{Escape string literals for inline SQL VALUES}
\usage{
db_escape_string_literal(conn, val)
}
\description{
Escape string literals for inline SQL VALUES
}
\keyword{internal}
