% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_generate_typed_values_sql}
\alias{db_generate_typed_values_sql}
\title{Generate type-aware VALUES SQL from data frame}
\usage{
db_generate_typed_values_sql(conn, data, col_types = NULL)
}
\description{
Generate type-aware VALUES SQL from data frame
}
\keyword{internal}
