% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{db_generate_typed_values_sql_for_view}
\alias{db_generate_typed_values_sql_for_view}
\title{Generate typed VALUES SQL for temporary views (helper)}
\usage{
db_generate_typed_values_sql_for_view(con, data)
}
\arguments{
\item{con}{DatabricksConnection object}

\item{data}{Data frame}
}
\value{
SQL VALUES clause
}
\description{
Generate typed VALUES SQL for temporary views (helper)
}
\keyword{internal}
