% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_generate_values_sql}
\alias{db_generate_values_sql}
\title{Generate VALUES SQL from data frame}
\usage{
db_generate_values_sql(conn, data)
}
\description{
Generate VALUES SQL from data frame
}
\keyword{internal}
