% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-init-scripts.R
\name{db_global_init_scripts_list}
\alias{db_global_init_scripts_list}
\title{Get All Global Init Scripts (Summaries)}
\usage{
db_global_init_scripts_list(
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Get a list of all global init scripts for this workspace.
}
\details{
This returns all properties for each script but not the script contents.
To retrieve the contents of a script, use \code{\link[=db_global_init_scripts_get]{db_global_init_scripts_get()}}
}
\seealso{
Other Global Init Script API: 
\code{\link{db_global_init_scripts_create}()},
\code{\link{db_global_init_scripts_delete}()},
\code{\link{db_global_init_scripts_get}()},
\code{\link{db_global_init_scripts_update}()}
}
\concept{Global Init Script API}
