% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-init-scripts.R
\name{db_global_init_scripts_update}
\alias{db_global_init_scripts_update}
\title{Update a Global Init Script}
\usage{
db_global_init_scripts_update(
  script_id,
  name = NULL,
  script = NULL,
  position = NULL,
  enabled = NULL,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{script_id}{String, the ID of the global init script.}

\item{name}{The name of the script}

\item{script}{String, content of script.}

\item{position}{Position of the global init script, must be >=1.
When not specified the script goes to last position.
When position already exists it will take position and displace existing.}

\item{enabled}{Boolean (default: false). Specifies whether the script is enabled.
The script runs only if enabled.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\description{
Update a Global Init Script
}
\details{
All fields are optional.
Unspecified fields retain their current value.
}
\seealso{
Other Global Init Script API: 
\code{\link{db_global_init_scripts_create}()},
\code{\link{db_global_init_scripts_delete}()},
\code{\link{db_global_init_scripts_get}()},
\code{\link{db_global_init_scripts_list}()}
}
\concept{Global Init Script API}
