% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{db_prepare_create_table_fields}
\alias{db_prepare_create_table_fields}
\title{Prepare fields for CREATE TABLE}
\usage{
db_prepare_create_table_fields(fields)
}
\description{
Prepare fields for CREATE TABLE
}
\keyword{internal}
