% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{db_query_get}
\alias{db_query_get}
\title{Get a SQL Query}
\usage{
db_query_get(id, host = db_host(), token = db_token(), perform_request = TRUE)
}
\arguments{
\item{id}{String, ID for the query.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
If \code{perform_request = TRUE}, returns endpoint-specific API output. If \code{FALSE}, returns an \code{httr2_request}.
}
\description{
Returns the repo with the given repo ID.
}
\seealso{
Other SQL Queries API: 
\code{\link{db_query_create}()},
\code{\link{db_query_delete}()},
\code{\link{db_query_list}()},
\code{\link{db_query_update}()}
}
\concept{SQL Queries API}
