% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-helpers.R
\name{db_req_error_body}
\alias{db_req_error_body}
\title{Propagate Databricks API Errors}
\usage{
db_req_error_body(resp)
}
\arguments{
\item{resp}{Object with class \code{httr2_response}.}
}
\description{
Propagate Databricks API Errors
}
\seealso{
Other Request Helpers: 
\code{\link{db_list_all_pages}()},
\code{\link{db_perform_request}()},
\code{\link{db_perform_response}()},
\code{\link{db_request}()},
\code{\link{db_request_json}()}
}
\concept{Request Helpers}
