% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{db_secrets_delete}
\alias{db_secrets_delete}
\title{Delete Secret in Secret Scope}
\usage{
db_secrets_delete(
  scope,
  key,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{scope}{Name of the scope that contains the secret to delete.}

\item{key}{Name of the secret to delete.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
If \code{perform_request = TRUE}, returns endpoint-specific API output. If \code{FALSE}, returns an \code{httr2_request}.
}
\description{
Delete Secret in Secret Scope
}
\details{
You must have \code{WRITE} or \code{MANAGE} permission on the secret scope.
\itemize{
\item Throws \code{RESOURCE_DOES_NOT_EXIST} if no such secret scope or secret exists.
\item Throws \code{PERMISSION_DENIED} if you do not have permission to make this API
call.
}
}
\seealso{
Other Secrets API: 
\code{\link{db_secrets_list}()},
\code{\link{db_secrets_put}()},
\code{\link{db_secrets_scope_acl_delete}()},
\code{\link{db_secrets_scope_acl_get}()},
\code{\link{db_secrets_scope_acl_list}()},
\code{\link{db_secrets_scope_acl_put}()},
\code{\link{db_secrets_scope_create}()},
\code{\link{db_secrets_scope_delete}()},
\code{\link{db_secrets_scope_list_all}()}
}
\concept{Secrets API}
