% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{db_sql_endpoint_start}
\alias{db_sql_endpoint_start}
\title{Start SQL Endpoint}
\usage{
db_sql_endpoint_start(
  id,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{id}{ID of the SQL endpoint.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Start SQL Endpoint
}
\seealso{
Other SQL Endpoints API: 
\code{\link{db_sql_endpoint_create}()},
\code{\link{db_sql_endpoint_delete}()},
\code{\link{db_sql_endpoint_edit}()},
\code{\link{db_sql_endpoint_get}()},
\code{\link{db_sql_endpoint_list}()},
\code{\link{db_sql_endpoint_stop}()},
\code{\link{db_sql_global_endpoint_edit}()},
\code{\link{db_sql_global_endpoint_get}()},
\code{\link{get_and_start_endpoint}()}
}
\concept{SQL Endpoints API}
