% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-query-execution.R
\name{db_sql_fetch_results_fast}
\alias{db_sql_fetch_results_fast}
\title{Fetch SQL Query Results (Fast Path)}
\usage{
db_sql_fetch_results_fast(
  resp,
  statement_id,
  manifest,
  return_arrow = FALSE,
  fetch_timeout = 300,
  row_limit = NULL,
  host = db_host(),
  token = db_token(),
  show_progress = TRUE
)
}
\description{
Fetch SQL Query Results (Fast Path)
}
\keyword{internal}
