% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-query-execution.R
\name{db_sql_fetch_results_parallel}
\alias{db_sql_fetch_results_parallel}
\title{Fetch SQL Query Results (Parallel Path)}
\usage{
db_sql_fetch_results_parallel(
  statement_id,
  manifest,
  last_chunk_index,
  return_arrow = FALSE,
  max_active_connections = 30,
  fetch_timeout = 300,
  row_limit = NULL,
  host = db_host(),
  token = db_token(),
  show_progress = TRUE
)
}
\description{
Fetch SQL Query Results (Parallel Path)
}
\keyword{internal}
