% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{db_sql_global_endpoint_edit}
\alias{db_sql_global_endpoint_edit}
\title{Edit Global SQL endpoint Config}
\usage{
db_sql_global_endpoint_edit(
  data_access_config = list(),
  sql_configuration_parameters = list(),
  instance_profile_arn = NULL,
  security_policy = "DATA_ACCESS_CONTROL",
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{data_access_config}{Named list of key-value pairs containing properties
for an external Hive metastore.}

\item{sql_configuration_parameters}{Named List of SQL configuration
parameters.}

\item{instance_profile_arn}{Instance profile used to access storage from SQL
endpoints.}

\item{security_policy}{The policy for controlling access to datasets. Must be
\code{DATA_ACCESS_CONTROL}.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Edit Global SQL endpoint Config
}
\details{
Important:
\itemize{
\item All fields are required.
\item Invoking this method restarts \strong{all} running SQL endpoints.
}
}
\seealso{
Other SQL Endpoints API: 
\code{\link{db_sql_endpoint_create}()},
\code{\link{db_sql_endpoint_delete}()},
\code{\link{db_sql_endpoint_edit}()},
\code{\link{db_sql_endpoint_get}()},
\code{\link{db_sql_endpoint_list}()},
\code{\link{db_sql_endpoint_start}()},
\code{\link{db_sql_endpoint_stop}()},
\code{\link{db_sql_global_endpoint_get}()},
\code{\link{get_and_start_endpoint}()}
}
\concept{SQL Endpoints API}
