% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{db_sql_global_warehouse_edit}
\alias{db_sql_global_warehouse_edit}
\title{Edit Global SQL warehouse Config}
\usage{
db_sql_global_warehouse_edit(
  data_access_config = list(),
  sql_configuration_parameters = list(),
  instance_profile_arn = NULL,
  security_policy = "DATA_ACCESS_CONTROL",
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{data_access_config}{Named list of key-value pairs containing properties
for an external Hive metastore.}

\item{sql_configuration_parameters}{Named List of SQL configuration
parameters.}

\item{instance_profile_arn}{Instance profile used to access storage from SQL
warehouses.}

\item{security_policy}{The policy for controlling access to datasets. Must be
\code{DATA_ACCESS_CONTROL}.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Edit Global SQL warehouse Config
}
\details{
Important:
\itemize{
\item All fields are required.
\item Invoking this method restarts \strong{all} running SQL warehouses.
}
}
\seealso{
Other SQL Endpoints API: 
\code{\link{db_sql_global_warehouse_get}()},
\code{\link{db_sql_warehouse_create}()},
\code{\link{db_sql_warehouse_delete}()},
\code{\link{db_sql_warehouse_edit}()},
\code{\link{db_sql_warehouse_get}()},
\code{\link{db_sql_warehouse_list}()},
\code{\link{db_sql_warehouse_start}()},
\code{\link{db_sql_warehouse_stop}()},
\code{\link{get_and_start_warehouse}()}
}
\concept{SQL Endpoints API}
