% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-query-execution.R
\name{db_sql_result}
\alias{db_sql_result}
\title{Get SQL Query Results}
\usage{
db_sql_result(
  statement_id,
  chunk_index,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{statement_id}{TODO}

\item{chunk_index}{TODO}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\description{
Get SQL Query Results
}
\details{
TODO
}
\seealso{
Other SQL Execution APIs: 
\code{\link{db_sql_cancel}()},
\code{\link{db_sql_query}()},
\code{\link{db_sql_status}()}
}
\concept{SQL Execution APIs}
