% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-catalogs.R
\name{db_uc_catalogs_get}
\alias{db_uc_catalogs_get}
\title{Get Catalog (Unity Catalog)}
\usage{
db_uc_catalogs_get(
  catalog,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{catalog}{The name of the catalog.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
List
}
\description{
Get Catalog (Unity Catalog)
}
\seealso{
Other Unity Catalog Management: 
\code{\link{db_uc_catalogs_list}()},
\code{\link{db_uc_schemas_get}()},
\code{\link{db_uc_schemas_list}()}
}
\concept{Unity Catalog Management}
