% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-volumes.R
\name{db_uc_volumes_create}
\alias{db_uc_volumes_create}
\title{Update Volume (Unity Catalog)}
\usage{
db_uc_volumes_create(
  catalog,
  schema,
  volume,
  volume_type = c("MANAGED", "EXTERNAL"),
  storage_location = NULL,
  comment = NULL,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{catalog}{Parent catalog of volume}

\item{schema}{Parent schema of volume}

\item{volume}{Volume name.}

\item{volume_type}{Either \code{'MANAGED'} or \code{'EXTERNAL'}.}

\item{storage_location}{The storage location on the cloud, only specified
when \code{volume_type} is \code{'EXTERNAL'}.}

\item{comment}{The comment attached to the volume.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
List
}
\description{
Update Volume (Unity Catalog)
}
\seealso{
Other Unity Catalog Volume Management: 
\code{\link{db_uc_volumes_delete}()},
\code{\link{db_uc_volumes_get}()},
\code{\link{db_uc_volumes_list}()},
\code{\link{db_uc_volumes_update}()}
}
\concept{Unity Catalog Volume Management}
