% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-fs.R
\name{db_volume_dir_create}
\alias{db_volume_dir_create}
\title{Volume FileSystem Create Directory}
\usage{
db_volume_dir_create(
  path,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{path}{Absolute path of the file in the Files API, omitting the initial
slash.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
If \code{perform_request = TRUE}, returns endpoint-specific API output. If \code{FALSE}, returns an \code{httr2_request}.
}
\description{
Volume FileSystem Create Directory
}
\seealso{
Other Volumes FileSystem API: 
\code{\link{db_volume_delete}()},
\code{\link{db_volume_dir_delete}()},
\code{\link{db_volume_dir_exists}()},
\code{\link{db_volume_download_dir}()},
\code{\link{db_volume_file_exists}()},
\code{\link{db_volume_list}()},
\code{\link{db_volume_read}()},
\code{\link{db_volume_upload_dir}()},
\code{\link{db_volume_write}()}
}
\concept{Volumes FileSystem API}
