% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-fs.R
\name{db_volume_list_files_recursive}
\alias{db_volume_list_files_recursive}
\title{Recursively collect file paths from a volume directory}
\usage{
db_volume_list_files_recursive(path, recurse = TRUE, host, token)
}
\description{
Recursively collect file paths from a volume directory
}
\keyword{internal}
