% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-search.R
\name{db_vs_indexes_scan}
\alias{db_vs_indexes_scan}
\title{Scan a Vector Search Index}
\usage{
db_vs_indexes_scan(
  endpoint,
  index,
  last_primary_key,
  num_results = 10,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{endpoint}{Name of vector search endpoint to scan}

\item{index}{Name of vector search index to scan}

\item{last_primary_key}{Primary key of the last entry returned in previous
scan}

\item{num_results}{Number of returns to return (default: 10)}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks bearer token or a \code{\link[=db_token_provider]{db_token_provider()}}, defaults to
calling \code{\link[=db_token]{db_token()}}. For a signed-in Shiny user's data, pass the
\code{token_provider()} returned by \code{\link[=db_shiny_server]{db_shiny_server()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
If \code{perform_request = TRUE}, returns endpoint-specific API output. If \code{FALSE}, returns an \code{httr2_request}.
}
\description{
Scan a Vector Search Index
}
\details{
Scan the specified vector index and return the first \code{num_results} entries
after the exclusive \code{primary_key}.
}
\seealso{
Other Vector Search API: 
\code{\link{db_vs_endpoints_create}()},
\code{\link{db_vs_endpoints_delete}()},
\code{\link{db_vs_endpoints_get}()},
\code{\link{db_vs_endpoints_list}()},
\code{\link{db_vs_indexes_create}()},
\code{\link{db_vs_indexes_delete}()},
\code{\link{db_vs_indexes_delete_data}()},
\code{\link{db_vs_indexes_get}()},
\code{\link{db_vs_indexes_list}()},
\code{\link{db_vs_indexes_query}()},
\code{\link{db_vs_indexes_query_next_page}()},
\code{\link{db_vs_indexes_sync}()},
\code{\link{db_vs_indexes_upsert_data}()},
\code{\link{delta_sync_index_spec}()},
\code{\link{direct_access_index_spec}()},
\code{\link{embedding_source_column}()},
\code{\link{embedding_vector_column}()}
}
\concept{Vector Search API}
