% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbBegin,DatabricksConnection-method}
\alias{dbBegin,DatabricksConnection-method}
\title{Begin transaction (not supported)}
\usage{
\S4method{dbBegin}{DatabricksConnection}(conn, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
Always throws an error (transactions not supported)
}
\description{
Begin transaction (not supported)
}
