% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbColumnInfo,DatabricksResult-method}
\alias{dbColumnInfo,DatabricksResult-method}
\title{Get column information from result}
\usage{
\S4method{dbColumnInfo}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
A data.frame with column names and types
}
\description{
Get column information from result
}
