% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbCommit,DatabricksConnection-method}
\alias{dbCommit,DatabricksConnection-method}
\title{Commit transaction (not supported)}
\usage{
\S4method{dbCommit}{DatabricksConnection}(conn, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
Always throws an error (read-only connection)
}
\description{
Commit transaction (not supported)
}
