% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbDataType,DatabricksConnection-method}
\alias{dbDataType,DatabricksConnection-method}
\title{Map R data types to Databricks SQL types}
\usage{
\S4method{dbDataType}{DatabricksConnection}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A DatabricksConnection object}

\item{obj}{R object(s) to get SQL types for}

\item{...}{Additional arguments (ignored)}
}
\value{
Character vector of SQL type names
}
\description{
Map R data types to Databricks SQL types
}
