% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbDisconnect,DatabricksConnection-method}
\alias{dbDisconnect,DatabricksConnection-method}
\title{Disconnect from Databricks}
\usage{
\S4method{dbDisconnect}{DatabricksConnection}(conn, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{TRUE} (invisibly)
}
\description{
Disconnect from Databricks
}
\details{
Invalidates this connection and its results, including copies of the
objects. Cancellation is requested for statements not known to be terminal.
}
