% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs_ls.R
\name{dbfs_ls}
\alias{dbfs_ls}
\title{List the contents of a path on DBFS}
\usage{
dbfs_ls(path, workspace, token = NULL, verbose = T, ...)
}
\arguments{
\item{path}{A string representing the path to list in DBFS}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used}

\item{verbose}{If TRUE, will print the API response to the console.  Defaults to
FALSE.}

\item{...}{Additional options to be passed}
}
\value{
The API response
}
\description{
List the contents of a directory, or the details of a file.  If the directory
does not exist a RESOURCE_DOES_NOT_EXIST exception will be thrown.
}
\details{
When calling list on a large directory, the list operation will time out after
 approximately 60s. We strongly recommend using list only on directories containing
 less than 10K files and discourage using the DBFS REST API for operations that
 list more than 10k files. Instead, we recommend that you perform such operations
 in the context of a cluster, using File system utilities, which provides the same
 functionality without timing out.

The API endpoint for listing a path on DBFS is '2.0/dbfs/list'.
  For all details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
# No need to include /dbfs/
path <- "/rk/data"

dbfs_ls(path = path, workspace = workspace)
}
