% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs_mkdirs.R
\name{dbfs_mkdirs}
\alias{dbfs_mkdirs}
\title{Create a path on DBFS}
\usage{
dbfs_mkdirs(path, workspace, token = NULL, verbose = T, ...)
}
\arguments{
\item{path}{A string representing the path to create in DBFS}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used}

\item{verbose}{If TRUE, will print the API response to the console.  Defaults to
FALSE.}

\item{...}{Additional options to be passed}
}
\value{
The API response
}
\description{
Create the given directory and necessary parent directories if they do not exist.
If there exists a file (not a directory) at any prefix of the input path, this
call throws an exception with RESOURCE_ALREADY_EXISTS. If this operation fails
it may have succeeded in creating some of the necessary parent directories.
}
\details{
The API endpoint for creating a path on DBFS is '2.0/dbfs/mkdirs'.
  For all details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
# No need to include /dbfs/
path <- "/rk/data/new_dir"

dbfs_mkdirs(path = path, workspace = workspace)
}
