% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs_mv.R
\name{dbfs_mv}
\alias{dbfs_mv}
\title{Move a file or directory on DBFS}
\usage{
dbfs_mv(from, to, workspace, token = NULL, verbose = T, ...)
}
\arguments{
\item{from}{A string representing the source path in DBFS}

\item{to}{A string representing the destination path in DBFS}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used}

\item{verbose}{If TRUE, will print the API response to the console.  Defaults to
FALSE.}

\item{...}{Additional options to be passed}
}
\value{
The API response
}
\description{
Move a file from one location to another location *within* DBFS. This will
not move files from your local system to DBFS.  See `dbfs_put()` and
`dbfs_read()` to work with your local filesystem.
}
\details{
If the source file does not exist, this call throws an exception with
RESOURCE_DOES_NOT_EXIST. If there already exists a file in the
destination path, this call throws an exception with
RESOURCE_ALREADY_EXISTS. If the given source path is a directory,
this call always recursively moves all files.

The API endpoint for moving files or directories on DBFS is '2.0/dbfs/move'.
  For all details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
# No need to include /dbfs/
source_path <- "/data/my_model.rds"
destination_path <- "/models/my_model.rds"

dbfs_mv(from = source_path, to = destination_path, workspace = workspace)
}
