% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs_put.R
\name{dbfs_put}
\alias{dbfs_put}
\title{Upload a file to DBFS}
\usage{
dbfs_put(
  local_file,
  destination_path,
  overwrite = "false",
  workspace,
  token = NULL,
  verbose = T,
  ...
)
}
\arguments{
\item{local_file}{Path to the file for upload}

\item{destination_path}{A string representing the destination path in DBFS}

\item{overwrite}{Boolean ('true', 'false') that specifies whether to overwrite existing files.
Defaults to 'false'}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used}

\item{verbose}{If TRUE, will print the API response to the console.  Defaults to
FALSE.}

\item{...}{Additional options to be passed}
}
\value{
The API response
}
\description{
Upload a file through the use of multipart form post. It is mainly
used for streaming uploads, but can also be used as a convenient
single call for data upload.
}
\details{
The API endpoint for putting files on DBFS is '2.0/dbfs/put'.
  For all details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
# No need to include /dbfs/
file <- "my_model.rds"
destination_path <- "/models/my_model.rds"

dbfs_put(file = file, destination_path = destination_path, workspace = workspace)
}
