% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfs_read.R
\name{dbfs_read}
\alias{dbfs_read}
\title{Read data from DBFS}
\usage{
dbfs_read(path, workspace, token = NULL, verbose = T, ...)
}
\arguments{
\item{path}{Absolute path to a file on DBFS}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used}

\item{verbose}{If TRUE, will print the API response to the console.  Defaults to
FALSE.}

\item{...}{Additional options to be passed}
}
\value{
The API response and the base6-encoded contents of the file read.
}
\description{
Return the contents of a file. If the file does not exist, this call
throws an exception with RESOURCE_DOES_NOT_EXIST.
}
\details{
The API endpoint for reading files from DBFS is '2.0/dbfs/read'.
  For all details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
# No need to include /dbfs/
file_path <- "/tmp/iris.json"

res <- dbfs_read(path = file_path, workspace = workspace)

data <- res$data

# Decode from base64
tidy_data <- jsonlite::fromJSON(rawToChar(jsonlite::base64_dec(data)))
}
