% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{dbfs_storage_info}
\alias{dbfs_storage_info}
\title{DBFS Storage Information}
\usage{
dbfs_storage_info(destination)
}
\arguments{
\item{destination}{DBFS destination. Example: \verb{dbfs:/my/path}.}
}
\description{
DBFS Storage Information
}
\seealso{
\code{\link[=cluster_log_conf]{cluster_log_conf()}}, \code{\link[=init_script_info]{init_script_info()}}

Other Cluster Log Configuration Objects: 
\code{\link{cluster_log_conf}()},
\code{\link{s3_storage_info}()}

Other Init Script Info Objects: 
\code{\link{file_storage_info}()},
\code{\link{s3_storage_info}()}
}
\concept{Cluster Log Configuration Objects}
\concept{Init Script Info Objects}
