% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbGetInfo,DatabricksConnection-method}
\alias{dbGetInfo,DatabricksConnection-method}
\title{Get connection information}
\usage{
\S4method{dbGetInfo}{DatabricksConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
A list with connection details
}
\description{
Get connection information
}
