% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbGetRowCount,DatabricksResult-method}
\alias{dbGetRowCount,DatabricksResult-method}
\title{Get number of rows fetched}
\usage{
\S4method{dbGetRowCount}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
Cumulative number of rows returned by successful fetches from this
result, shared across copies of the result object.
}
\description{
Get number of rows fetched
}
