% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbGetRowsAffected,DatabricksResult-method}
\alias{dbGetRowsAffected,DatabricksResult-method}
\title{Get number of rows affected by a statement}
\usage{
\S4method{dbGetRowsAffected}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
The reported affected row count for statement results, or zero when
no affected-row field is returned. Query results return \code{-1}.
}
\description{
Get number of rows affected by a statement
}
