% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbGetStatement,DatabricksResult-method}
\alias{dbGetStatement,DatabricksResult-method}
\title{Get SQL statement from result}
\usage{
\S4method{dbGetStatement}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
The SQL statement as character
}
\description{
Get SQL statement from result
}
