% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbIsValid,DatabricksConnection-method}
\alias{dbIsValid,DatabricksConnection-method}
\title{Check if connection is valid}
\usage{
\S4method{dbIsValid}{DatabricksConnection}(dbObj, ...)
}
\arguments{
\item{dbObj}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{TRUE} if connection is valid, \code{FALSE} otherwise
}
\description{
Check if connection is valid
}
