% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbIsValid,DatabricksResult-method}
\alias{dbIsValid,DatabricksResult-method}
\title{Check whether a Databricks result is valid}
\usage{
\S4method{dbIsValid}{DatabricksResult}(dbObj, ...)
}
\arguments{
\item{dbObj}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{TRUE} until the result is cleared or its connection is disconnected.
}
\description{
Check whether a Databricks result is valid
}
\concept{DBI Backend}
