% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbListFields,DatabricksConnection,AsIs-method}
\alias{dbListFields,DatabricksConnection,AsIs-method}
\title{List column names of a Databricks table (AsIs method)}
\usage{
\S4method{dbListFields}{DatabricksConnection,AsIs}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name as AsIs object (from I())}

\item{...}{Additional arguments (ignored)}
}
\value{
Character vector of column names
}
\description{
List column names of a Databricks table (AsIs method)
}
