% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbListFields,DatabricksConnection,character-method}
\alias{dbListFields,DatabricksConnection,character-method}
\title{List column names of a Databricks table}
\usage{
\S4method{dbListFields}{DatabricksConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name to describe}

\item{...}{Additional arguments (ignored)}
}
\value{
Character vector of column names
}
\description{
List column names of a Databricks table
}
