% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbListTables,DatabricksConnection-method}
\alias{dbListTables,DatabricksConnection-method}
\title{List tables in Databricks catalog/schema}
\usage{
\S4method{dbListTables}{DatabricksConnection}(conn, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
Character vector of table names
}
\description{
List tables in Databricks catalog/schema
}
