% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{dbplyr_edition.DatabricksConnection}
\alias{dbplyr_edition.DatabricksConnection}
\title{Declare dbplyr API version for Databricks connections}
\usage{
\method{dbplyr_edition}{DatabricksConnection}(con)
}
\arguments{
\item{con}{A DatabricksConnection object}
}
\value{
The dbplyr edition number (2L)
}
\description{
Declare dbplyr API version for Databricks connections
}
