% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbQuoteIdentifier,DatabricksConnection,character-method}
\alias{dbQuoteIdentifier,DatabricksConnection,character-method}
\title{Quote identifiers for Databricks SQL}
\usage{
\S4method{dbQuoteIdentifier}{DatabricksConnection,character}(conn, x, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{x}{Character vector of identifiers to quote}

\item{...}{Additional arguments (ignored)}
}
\value{
SQL object with quoted identifiers
}
\description{
Quote identifiers for Databricks SQL
}
