% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{dbQuoteIdentifier,DatabricksConnection,SQL-method}
\alias{dbQuoteIdentifier,DatabricksConnection,SQL-method}
\title{Quote SQL objects (passthrough)}
\usage{
\S4method{dbQuoteIdentifier}{DatabricksConnection,SQL}(conn, x, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{x}{SQL object (already quoted)}

\item{...}{Additional arguments (ignored)}
}
\value{
The SQL object unchanged
}
\description{
Quote SQL objects (passthrough)
}
