% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbReadTable,DatabricksConnection,AsIs-method}
\alias{dbReadTable,DatabricksConnection,AsIs-method}
\title{Read a Databricks table (AsIs method)}
\usage{
\S4method{dbReadTable}{DatabricksConnection,AsIs}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name as AsIs object (from I())}

\item{...}{Additional arguments passed to dbGetQuery}
}
\value{
A data.frame with table contents
}
\description{
Read a Databricks table (AsIs method)
}
