% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbReadTable,DatabricksConnection,character-method}
\alias{dbReadTable,DatabricksConnection,character-method}
\title{Read a Databricks table}
\usage{
\S4method{dbReadTable}{DatabricksConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name to read}

\item{...}{Additional arguments passed to dbGetQuery}
}
\value{
A data.frame with table contents
}
\description{
Read a Databricks table
}
