% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbReadTable,DatabricksConnection,Id-method}
\alias{dbReadTable,DatabricksConnection,Id-method}
\title{Read a Databricks table (Id method)}
\usage{
\S4method{dbReadTable}{DatabricksConnection,Id}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name as Id object}

\item{...}{Additional arguments passed to dbGetQuery}
}
\value{
A data.frame with table contents
}
\description{
Read a Databricks table (Id method)
}
