% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbRemoveTable,DatabricksConnection,AsIs-method}
\alias{dbRemoveTable,DatabricksConnection,AsIs-method}
\title{Remove a Databricks table (AsIs method)}
\usage{
\S4method{dbRemoveTable}{DatabricksConnection,AsIs}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name as AsIs object (from I())}

\item{...}{Additional arguments (ignored)}
}
\value{
TRUE invisibly on success
}
\description{
Remove a Databricks table (AsIs method)
}
