% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbRemoveTable,DatabricksConnection,character-method}
\alias{dbRemoveTable,DatabricksConnection,character-method}
\title{Remove a Databricks table}
\usage{
\S4method{dbRemoveTable}{DatabricksConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name to remove}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{TRUE} invisibly on success
}
\description{
Remove a Databricks table
}
