% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbRemoveTable,DatabricksConnection,Id-method}
\alias{dbRemoveTable,DatabricksConnection,Id-method}
\title{Remove a Databricks table (Id method)}
\usage{
\S4method{dbRemoveTable}{DatabricksConnection,Id}(conn, name, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{name}{Table name as Id object}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{TRUE} invisibly on success
}
\description{
Remove a Databricks table (Id method)
}
