% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbRollback,DatabricksConnection-method}
\alias{dbRollback,DatabricksConnection-method}
\title{Rollback transaction (not supported)}
\usage{
\S4method{dbRollback}{DatabricksConnection}(conn, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
Always throws an error (read-only connection)
}
\description{
Rollback transaction (not supported)
}
