% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbSendQuery,DatabricksConnection,character-method}
\alias{dbSendQuery,DatabricksConnection,character-method}
\title{Send query to Databricks (asynchronous)}
\usage{
\S4method{dbSendQuery}{DatabricksConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{statement}{SQL statement to execute}

\item{...}{Additional arguments (ignored)}
}
\value{
A DatabricksResult object
}
\description{
Send query to Databricks (asynchronous)
}
