% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_job.R
\name{delete_job}
\alias{delete_job}
\title{Delete a job on Databricks}
\usage{
delete_job(job_id = NULL, name = NULL, workspace, token = NULL, verbose = T)
}
\arguments{
\item{job_id}{A number representing the unique identifier of a job on Databricks.}

\item{name}{Optional.  A string representing the name of the job.  If multiple
jobs share the same name, you'll need to rename the jobs or provide
the unique job ID.}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided, netrc will be used.}

\item{verbose}{If true, will pretty print the success or failure of the
request and add a `job_id` variable to the R environment.  Defaults to TRUE.}
}
\value{
The API response.
}
\description{
Delete the job and send an email to the addresses specified in
JobSettings.email_notifications. No action occurs if the job
has already been removed. After the job is removed, neither
 its details or its run history is visible via the Jobs UI or API.
}
\details{
The job is guaranteed to be removed upon completion of this request.
 However, runs that were active before the receipt of this request
 may still be active. They will be terminated asynchronously.
The API endpoint for deleting a job is '2.0/jobs/delete'.  For all details
on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
delete_job(job_id = 206,
workspace = "https://dbc-z64b06b4-d212.cloud.databricks.com",
token = "dapi310240980a9dgqwebdsfadsf21")

}
