% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-helpers.R
\name{detect_brickster_venv}
\alias{detect_brickster_venv}
\title{Detect brickster virtualenv}
\usage{
detect_brickster_venv()
}
\description{
Detect brickster virtualenv
}
\details{
Returns \code{NULL} when running within Databricks,
otherwise "r-brickster"
}
