% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{docker_image}
\alias{docker_image}
\title{Docker Image}
\usage{
docker_image(url, username, password)
}
\arguments{
\item{url}{URL for the Docker image.}

\item{username}{User name for the Docker repository.}

\item{password}{Password for the Docker repository.}
}
\description{
Docker image connection information.
}
\details{
Uses basic authentication, \strong{strongly} recommended that credentials are not
stored in any scripts and environment variables should be used.
}
\seealso{
\code{\link[=db_cluster_create]{db_cluster_create()}}, \code{\link[=db_cluster_edit]{db_cluster_edit()}}
}
