% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{embedding_source_column}
\alias{embedding_source_column}
\title{Embedding Source Column}
\usage{
embedding_source_column(name, model_endpoint_name)
}
\arguments{
\item{name}{Name of the column}

\item{model_endpoint_name}{Name of the embedding model endpoint}
}
\description{
Embedding Source Column
}
\seealso{
Other Vector Search API: 
\code{\link{db_vs_endpoints_create}()},
\code{\link{db_vs_endpoints_delete}()},
\code{\link{db_vs_endpoints_get}()},
\code{\link{db_vs_endpoints_list}()},
\code{\link{db_vs_indexes_create}()},
\code{\link{db_vs_indexes_delete}()},
\code{\link{db_vs_indexes_delete_data}()},
\code{\link{db_vs_indexes_get}()},
\code{\link{db_vs_indexes_list}()},
\code{\link{db_vs_indexes_query}()},
\code{\link{db_vs_indexes_query_next_page}()},
\code{\link{db_vs_indexes_scan}()},
\code{\link{db_vs_indexes_sync}()},
\code{\link{db_vs_indexes_upsert_data}()},
\code{\link{delta_sync_index_spec}()},
\code{\link{direct_access_index_spec}()},
\code{\link{embedding_vector_column}()}
}
\concept{Vector Search API}
