% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{file_storage_info}
\alias{file_storage_info}
\title{File Storage Information}
\usage{
file_storage_info(destination)
}
\arguments{
\item{destination}{File destination. Example: \verb{file:/my/file.sh}.}
}
\description{
File Storage Information
}
\details{
The file storage type is only available for clusters set up using Databricks
Container Services.
}
\seealso{
\code{\link[=init_script_info]{init_script_info()}}

Other Init Script Info Objects: 
\code{\link{dbfs_storage_info}()},
\code{\link{s3_storage_info}()}
}
\concept{Init Script Info Objects}
