% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{for_each_task}
\alias{for_each_task}
\title{For Each Task}
\usage{
for_each_task(inputs, task, concurrency = 1)
}
\arguments{
\item{inputs}{Array for task to iterate on. This can be a JSON string or a
reference to an array parameter.}

\item{task}{Must be a \code{\link[=job_task]{job_task()}}.}

\item{concurrency}{Maximum allowed number of concurrent runs of the task.}
}
\description{
For Each Task
}
\seealso{
Other Task Objects: 
\code{\link{condition_task}()},
\code{\link{email_notifications}()},
\code{\link{libraries}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{pipeline_task}()},
\code{\link{python_wheel_task}()},
\code{\link{run_job_task}()},
\code{\link{spark_jar_task}()},
\code{\link{spark_python_task}()},
\code{\link{spark_submit_task}()},
\code{\link{sql_file_task}()},
\code{\link{sql_query_task}()}
}
\concept{Task Objects}
