% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{gcp_attributes}
\alias{gcp_attributes}
\title{GCP Attributes}
\usage{
gcp_attributes(use_preemptible_executors = TRUE, google_service_account = NULL)
}
\arguments{
\item{use_preemptible_executors}{Boolean (Default: \code{TRUE}). If \code{TRUE} Uses
preemptible executors}

\item{google_service_account}{Google service account email address that the
cluster uses to authenticate with Google Identity. This field is used for
authentication with the GCS and BigQuery data sources.}
}
\description{
GCP Attributes
}
\details{
For use with GCS and BigQuery, your Google service account that you use to
access the data source must be in the same project as the SA that you
specified when setting up your Databricks account.
}
\seealso{
\code{\link[=db_cluster_create]{db_cluster_create()}}, \code{\link[=db_cluster_edit]{db_cluster_edit()}}

Other Cloud Attributes: 
\code{\link{aws_attributes}()},
\code{\link{azure_attributes}()}
}
\concept{Cloud Attributes}
