% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{generate_temp_name}
\alias{generate_temp_name}
\title{Generate unique temporary table/view name}
\usage{
generate_temp_name(prefix = "dbplyr_temp")
}
\arguments{
\item{prefix}{Base name prefix (default: "dbplyr_temp")}
}
\value{
Unique temporary name
}
\description{
Generate unique temporary table/view name
}
\keyword{internal}
