% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_and_start_endpoint}
\alias{get_and_start_endpoint}
\title{Get and Start Endpoint}
\usage{
get_and_start_endpoint(
  id,
  polling_interval = 5,
  host = db_host(),
  token = db_token()
)
}
\arguments{
\item{id}{ID of the SQL endpoint.}

\item{polling_interval}{Number of seconds to wait between status checks}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\value{
\code{db_sql_endpoint_get()}
}
\description{
Get and Start Endpoint
}
\details{
Get information regarding a Databricks cluster. If the cluster is
inactive it will be started and wait until the cluster is active.
}
\seealso{
\code{\link[=db_sql_endpoint_get]{db_sql_endpoint_get()}} and \code{\link[=db_sql_endpoint_start]{db_sql_endpoint_start()}}.

Other SQL Endpoints API: 
\code{\link{db_sql_endpoint_create}()},
\code{\link{db_sql_endpoint_delete}()},
\code{\link{db_sql_endpoint_edit}()},
\code{\link{db_sql_endpoint_get}()},
\code{\link{db_sql_endpoint_list}()},
\code{\link{db_sql_endpoint_start}()},
\code{\link{db_sql_endpoint_stop}()},
\code{\link{db_sql_global_endpoint_edit}()},
\code{\link{db_sql_global_endpoint_get}()}
}
\concept{Endpoint Helpers}
\concept{SQL Endpoints API}
