% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_status.R
\name{get_cluster_status}
\alias{get_cluster_status}
\title{Retrieve the information for a cluster.}
\usage{
get_cluster_status(cluster_id, workspace, token = NULL, verbose = T, ...)
}
\arguments{
\item{cluster_id}{A string containing the unique id for an online Databricks cluster}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{verbose}{If TRUE, will print the API response to the console.  Defaults to
FALSE.}

\item{...}{Additional options to be passed to \code{data.table::fread} which is used to
parse the API response.}
}
\value{
A list with four elements:
\itemize{
    \item \emph{response} - The full API response, including status code.
    \item \emph{response_json} - The JSON result.
    \item \emph{response_list} - The JSON result transformed into a list.
    \item \emph{response_df} - The JSON result transformed into a dataframe.
}
}
\description{
Will retrieve detailed information on a Databricks cluster.  You can locate the cluster ID in
 the URL of the cluster configuration page.  For example:
}
\details{
https://mycompany.cloud.databricks.com/#/setting/clusters/xxxx-xxxxx-xxxxxx/

Where xxxx-xxxxx-xxxxxx is the cluster ID.

The API endpoint for terminating a cluster is '2.0/clusters/get'.
  For all details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
workspace <- "https://myworkspace.cloud.databricks.com"
cluster_id <- "0818-155203-cheese22"

get_cluster_status(workspace = workspace, cluster_id = cluster_id)
}
