% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_status.R
\name{get_run_status}
\alias{get_run_status}
\title{Get the status of a job run on Databricks}
\usage{
get_run_status(run_id, workspace, token = NULL, verbose = TRUE)
}
\arguments{
\item{run_id}{A number representing a valid run id generating by launching
a job on Databricks.}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used.}

\item{verbose}{If true, will pretty print the success or failure of the
request and add a run_id variable to the R environment.  Defaults to TRUE.}
}
\value{
A list with four elements:
\itemize{
    \item \emph{response} - The full API response, including status code.
    \item \emph{response_json} - The JSON result.
    \item \emph{response_list} - The JSON result transformed into a list.
    \item \emph{response_df} - The JSON result transformed into a dataframe.
}
}
\description{
Fetches the status of a job run on Databricks, whether it has completed or
not.  Includes details about start time, cluster spec, task, and more.
}
\details{
The API endpoint for running a job is '2.0/jobs/runs/get'.  For all
details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
res <- get_run_status(run_id = 2932371,
                      workspace = "https://eastus2.azuredatabricks.net",
                      token = "dapiafoi32984320890d9asaf1")

## Get data in different formats
resjson <- res$response_json
reslist <- res$response_list
res_df <- res$response_df

}
