% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-helpers.R
\name{in_databricks_nb}
\alias{in_databricks_nb}
\title{Detect if running within Databricks Notebook}
\usage{
in_databricks_nb()
}
\value{
Boolean
}
\description{
Detect if running within Databricks Notebook
}
\details{
R sessions on Databricks can be detected via various environment variables
and directories.
}
