% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{init_script_info}
\alias{init_script_info}
\title{Init Script Info}
\usage{
init_script_info(...)
}
\arguments{
\item{...}{Accepts multiple instances \code{\link[=s3_storage_info]{s3_storage_info()}},
\code{\link[=file_storage_info]{file_storage_info()}}, or \code{\link[=dbfs_storage_info]{dbfs_storage_info()}}.}
}
\description{
Init Script Info
}
\details{
\code{\link[=file_storage_info]{file_storage_info()}} is only available for clusters set up using Databricks
Container Services.

For instructions on using init scripts with Databricks Container Services,
see \href{https://docs.databricks.com/clusters/custom-containers.html#containers-init-script}{Use an init script}.
}
\seealso{
\code{\link[=db_cluster_create]{db_cluster_create()}}, \code{\link[=db_cluster_edit]{db_cluster_edit()}}
}
