% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.access_control_req_group}
\alias{is.access_control_req_group}
\title{Test if object is of class AccessControlRequestForGroup}
\usage{
is.access_control_req_group(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{AccessControlRequestForGroup}
class.
}
\description{
Test if object is of class AccessControlRequestForGroup
}
