% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.access_control_req_user}
\alias{is.access_control_req_user}
\title{Test if object is of class AccessControlRequestForUser}
\usage{
is.access_control_req_user(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{AccessControlRequestForUser}
class.
}
\description{
Test if object is of class AccessControlRequestForUser
}
