% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.access_control_request}
\alias{is.access_control_request}
\title{Test if object is of class AccessControlRequest}
\usage{
is.access_control_request(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{AccessControlRequest} class.
}
\description{
Test if object is of class AccessControlRequest
}
