% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.aws_attributes}
\alias{is.aws_attributes}
\title{Test if object is of class AwsAttributes}
\usage{
is.aws_attributes(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{AwsAttributes} class.
}
\description{
Test if object is of class AwsAttributes
}
