% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.azure_attributes}
\alias{is.azure_attributes}
\title{Test if object is of class AzureAttributes}
\usage{
is.azure_attributes(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{AzureAttributes} class.
}
\description{
Test if object is of class AzureAttributes
}
