% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.cluster_autoscale}
\alias{is.cluster_autoscale}
\title{Test if object is of class AutoScale}
\usage{
is.cluster_autoscale(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{AutoScale} class.
}
\description{
Test if object is of class AutoScale
}
