% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.cluster_log_conf}
\alias{is.cluster_log_conf}
\title{Test if object is of class ClusterLogConf}
\usage{
is.cluster_log_conf(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{ClusterLogConf} class.
}
\description{
Test if object is of class ClusterLogConf
}
