% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.condition_task}
\alias{is.condition_task}
\title{Test if object is of class ConditionTask}
\usage{
is.condition_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{ConditionTask} class.
}
\description{
Test if object is of class ConditionTask
}
