% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.cron_schedule}
\alias{is.cron_schedule}
\title{Test if object is of class CronSchedule}
\usage{
is.cron_schedule(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{CronSchedule} class.
}
\description{
Test if object is of class CronSchedule
}
