% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.delta_sync_index}
\alias{is.delta_sync_index}
\title{Test if object is of class DeltaSyncIndex}
\usage{
is.delta_sync_index(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{DeltaSyncIndex} class.
}
\description{
Test if object is of class DeltaSyncIndex
}
