% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.direct_access_index}
\alias{is.direct_access_index}
\title{Test if object is of class DirectAccessIndex}
\usage{
is.direct_access_index(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{DirectAccessIndex} class.
}
\description{
Test if object is of class DirectAccessIndex
}
