% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.docker_image}
\alias{is.docker_image}
\title{Test if object is of class DockerImage}
\usage{
is.docker_image(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{DockerImage} class.
}
\description{
Test if object is of class DockerImage
}
