% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.email_notifications}
\alias{is.email_notifications}
\title{Test if object is of class JobEmailNotifications}
\usage{
is.email_notifications(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{JobEmailNotifications} class.
}
\description{
Test if object is of class JobEmailNotifications
}
