% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.embedding_source_column}
\alias{is.embedding_source_column}
\title{Test if object is of class EmbeddingSourceColumn}
\usage{
is.embedding_source_column(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{EmbeddingSourceColumn} class.
}
\description{
Test if object is of class EmbeddingSourceColumn
}
