% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.embedding_vector_column}
\alias{is.embedding_vector_column}
\title{Test if object is of class EmbeddingVectorColumn}
\usage{
is.embedding_vector_column(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{EmbeddingVectorColumn} class.
}
\description{
Test if object is of class EmbeddingVectorColumn
}
