% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.file_storage_info}
\alias{is.file_storage_info}
\title{Test if object is of class FileStorageInfo}
\usage{
is.file_storage_info(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{FileStorageInfo} class.
}
\description{
Test if object is of class FileStorageInfo
}
