% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.for_each_task}
\alias{is.for_each_task}
\title{Test if object is of class ForEachTask}
\usage{
is.for_each_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{ForEachTask} class.
}
\description{
Test if object is of class ForEachTask
}
