% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.gcp_attributes}
\alias{is.gcp_attributes}
\title{Test if object is of class GcpAttributes}
\usage{
is.gcp_attributes(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{GcpAttributes} class.
}
\description{
Test if object is of class GcpAttributes
}
