% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.git_source}
\alias{is.git_source}
\title{Test if object is of class GitSource}
\usage{
is.git_source(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{GitSource}
class.
}
\description{
Test if object is of class GitSource
}
