% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.init_script_info}
\alias{is.init_script_info}
\title{Test if object is of class InitScriptInfo}
\usage{
is.init_script_info(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{InitScriptInfo} class.
}
\description{
Test if object is of class InitScriptInfo
}
