% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.job_task}
\alias{is.job_task}
\title{Test if object is of class JobTaskSettings}
\usage{
is.job_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{JobTaskSettings} class.
}
\description{
Test if object is of class JobTaskSettings
}
