% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.lib_cran}
\alias{is.lib_cran}
\title{Test if object is of class CranLibrary}
\usage{
is.lib_cran(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{CranLibrary} class.
}
\description{
Test if object is of class CranLibrary
}
