% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.lib_egg}
\alias{is.lib_egg}
\title{Test if object is of class EggLibrary}
\usage{
is.lib_egg(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{EggLibrary} class.
}
\description{
Test if object is of class EggLibrary
}
