% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.lib_jar}
\alias{is.lib_jar}
\title{Test if object is of class JarLibrary}
\usage{
is.lib_jar(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{JarLibrary} class.
}
\description{
Test if object is of class JarLibrary
}
