% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.lib_maven}
\alias{is.lib_maven}
\title{Test if object is of class MavenLibrary}
\usage{
is.lib_maven(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{MavenLibrary} class.
}
\description{
Test if object is of class MavenLibrary
}
