% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.lib_pypi}
\alias{is.lib_pypi}
\title{Test if object is of class PyPiLibrary}
\usage{
is.lib_pypi(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{PyPiLibrary} class.
}
\description{
Test if object is of class PyPiLibrary
}
