% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.lib_whl}
\alias{is.lib_whl}
\title{Test if object is of class WhlLibrary}
\usage{
is.lib_whl(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{WhlLibrary} class.
}
\description{
Test if object is of class WhlLibrary
}
