% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.libraries}
\alias{is.libraries}
\title{Test if object is of class Libraries}
\usage{
is.libraries(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{Libraries} class.
}
\description{
Test if object is of class Libraries
}
