% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.library}
\alias{is.library}
\title{Test if object is of class Library}
\usage{
is.library(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{Library} class.
}
\description{
Test if object is of class Library
}
