% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.new_cluster}
\alias{is.new_cluster}
\title{Test if object is of class NewCluster}
\usage{
is.new_cluster(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{NewCluster} class.
}
\description{
Test if object is of class NewCluster
}
