% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.notebook_task}
\alias{is.notebook_task}
\title{Test if object is of class NotebookTask}
\usage{
is.notebook_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{NotebookTask} class.
}
\description{
Test if object is of class NotebookTask
}
