% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.pipeline_task}
\alias{is.pipeline_task}
\title{Test if object is of class PipelineTask}
\usage{
is.pipeline_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{PipelineTask} class.
}
\description{
Test if object is of class PipelineTask
}
