% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.python_wheel_task}
\alias{is.python_wheel_task}
\title{Test if object is of class PythonWheelTask}
\usage{
is.python_wheel_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{PythonWheelTask} class.
}
\description{
Test if object is of class PythonWheelTask
}
