% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.run_job_task}
\alias{is.run_job_task}
\title{Test if object is of class RunJobTask}
\usage{
is.run_job_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{RunJobTask} class.
}
\description{
Test if object is of class RunJobTask
}
