% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.s3_storage_info}
\alias{is.s3_storage_info}
\title{Test if object is of class S3StorageInfo}
\usage{
is.s3_storage_info(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{S3StorageInfo} class.
}
\description{
Test if object is of class S3StorageInfo
}
