% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.spark_jar_task}
\alias{is.spark_jar_task}
\title{Test if object is of class SparkJarTask}
\usage{
is.spark_jar_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{SparkJarTask} class.
}
\description{
Test if object is of class SparkJarTask
}
