% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.spark_python_task}
\alias{is.spark_python_task}
\title{Test if object is of class SparkPythonTask}
\usage{
is.spark_python_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{SparkPythonTask} class.
}
\description{
Test if object is of class SparkPythonTask
}
