% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.sql_file_task}
\alias{is.sql_file_task}
\title{Test if object is of class SqlFileTask}
\usage{
is.sql_file_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{SqlFileTask} class.
}
\description{
Test if object is of class SqlFileTask
}
