% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.sql_query_task}
\alias{is.sql_query_task}
\title{Test if object is of class SqlQueryTask}
\usage{
is.sql_query_task(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{SqlQueryTask} class.
}
\description{
Test if object is of class SqlQueryTask
}
