% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.valid_task_type}
\alias{is.valid_task_type}
\title{Test if object is of class JobTask}
\usage{
is.valid_task_type(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{JobTask} class.
}
\description{
Test if object is of class JobTask
}
