% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{is.vector_search_index_spec}
\alias{is.vector_search_index_spec}
\title{Test if object is of class VectorSearchIndexSpec}
\usage{
is.vector_search_index_spec(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{VectorSearchIndexSpec} class.
}
\description{
Test if object is of class VectorSearchIndexSpec
}
