% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_list.R
\name{jobs_list}
\alias{jobs_list}
\title{List all jobs in a Databricks workspace.}
\usage{
jobs_list(workspace, token = NULL, verbose = T)
}
\arguments{
\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used.}

\item{verbose}{If true, will pretty print the success or failure of the
request.  Defaults to TRUE.}
}
\value{
A list with four elements:
\itemize{
    \item \emph{response} - The full API response, including status code.
    \item \emph{response_list} - The JSON result transformed into a list.
    \item \emph{response_df} - The JSON result transformed into a dataframe.
    \item \emph{response_tidy} - A simple dataframe with the job ID, name,
    time of creation, and who created it.
}
If the details of jobs are not needed, use the \emph{response_tidy}
dataframe.
}
\description{
Fetches all of the details for jobs in the authenticated workspace.
See also \code{runs_list} to learn more about a particular job run.
}
\details{
The API endpoint for running a job is '2.0/jobs/list'.  For all
details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
jobs <- jobs_list(workspace = "https://eastus2.azuredatabricks.net",
                  token = "dapiafoi32984320890d9asaf1")

## Get data in different formats
jobs_json <- jobs$response_json
jobs_list <- jobs$response_list
jobs_df <- jobs$response_df

}
