% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{lib_cran}
\alias{lib_cran}
\title{Cran Library (R)}
\usage{
lib_cran(package, repo = NULL)
}
\arguments{
\item{package}{The name of the CRAN package to install.}

\item{repo}{The repository where the package can be found. If not specified,
the default CRAN repo is used.}
}
\description{
Cran Library (R)
}
\seealso{
\code{\link[=libraries]{libraries()}}

Other Library Objects: 
\code{\link{lib_egg}()},
\code{\link{lib_jar}()},
\code{\link{lib_maven}()},
\code{\link{lib_pypi}()},
\code{\link{lib_whl}()},
\code{\link{libraries}()}
}
\concept{Library Objects}
