% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{lib_egg}
\alias{lib_egg}
\title{Egg Library (Python)}
\usage{
lib_egg(egg)
}
\arguments{
\item{egg}{URI of the egg to be installed. DBFS and S3 URIs are supported.
For example: \verb{dbfs:/my/egg} or \verb{s3://my-bucket/egg}. If S3 is used, make sure
the cluster has read access on the library. You may need to launch the
cluster with an instance profile to access the S3 URI.}
}
\description{
Egg Library (Python)
}
\seealso{
\code{\link[=libraries]{libraries()}}

Other Library Objects: 
\code{\link{lib_cran}()},
\code{\link{lib_jar}()},
\code{\link{lib_maven}()},
\code{\link{lib_pypi}()},
\code{\link{lib_whl}()},
\code{\link{libraries}()}
}
\concept{Library Objects}
