% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{lib_jar}
\alias{lib_jar}
\title{Jar Library (Scala)}
\usage{
lib_jar(jar)
}
\arguments{
\item{jar}{URI of the JAR to be installed. DBFS and S3 URIs are supported.
For example: \verb{dbfs:/mnt/databricks/library.jar} or
\verb{s3://my-bucket/library.jar}. If S3 is used, make sure the cluster has read
access on the library. You may need to launch the cluster with an instance
profile to access the S3 URI.}
}
\description{
Jar Library (Scala)
}
\seealso{
\code{\link[=libraries]{libraries()}}

Other Library Objects: 
\code{\link{lib_cran}()},
\code{\link{lib_egg}()},
\code{\link{lib_maven}()},
\code{\link{lib_pypi}()},
\code{\link{lib_whl}()},
\code{\link{libraries}()}
}
\concept{Library Objects}
