% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{lib_maven}
\alias{lib_maven}
\title{Maven Library (Scala)}
\usage{
lib_maven(coordinates, repo = NULL, exclusions = NULL)
}
\arguments{
\item{coordinates}{Gradle-style Maven coordinates. For example:
\verb{org.jsoup:jsoup:1.7.2}.}

\item{repo}{Maven repo to install the Maven package from. If omitted, both
Maven Central Repository and Spark Packages are searched.}

\item{exclusions}{List of dependencies to exclude. For example:
\code{list("slf4j:slf4j", "*:hadoop-client")}.
\href{https://maven.apache.org/guides/introduction/introduction-to-optional-and-excludes-dependencies.html}{Maven dependency exclusions}.}
}
\description{
Maven Library (Scala)
}
\seealso{
\code{\link[=libraries]{libraries()}}

Other Library Objects: 
\code{\link{lib_cran}()},
\code{\link{lib_egg}()},
\code{\link{lib_jar}()},
\code{\link{lib_pypi}()},
\code{\link{lib_whl}()},
\code{\link{libraries}()}
}
\concept{Library Objects}
