% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{lib_pypi}
\alias{lib_pypi}
\title{PyPi Library (Python)}
\usage{
lib_pypi(package, repo = NULL)
}
\arguments{
\item{package}{The name of the PyPI package to install. An optional exact
version specification is also supported. Examples: \code{simplejson} and
\verb{simplejson==3.8.0}.}

\item{repo}{The repository where the package can be found. If not specified,
the default pip index is used.}
}
\description{
PyPi Library (Python)
}
\seealso{
\code{\link[=libraries]{libraries()}}

Other Library Objects: 
\code{\link{lib_cran}()},
\code{\link{lib_egg}()},
\code{\link{lib_jar}()},
\code{\link{lib_maven}()},
\code{\link{lib_whl}()},
\code{\link{libraries}()}
}
\concept{Library Objects}
