% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{lib_whl}
\alias{lib_whl}
\title{Wheel Library (Python)}
\usage{
lib_whl(whl)
}
\arguments{
\item{whl}{URI of the wheel or zipped wheels to be installed.
DBFS and S3 URIs are supported. For example: \verb{dbfs:/my/whl} or
\verb{s3://my-bucket/whl}. If S3 is used, make sure the cluster has read access on
the library. You may need to launch the cluster with an instance profile to
access the S3 URI. Also the wheel file name needs to use the correct
convention. If zipped wheels are to be installed, the file name suffix should
be \code{.wheelhouse.zip}.}
}
\description{
Wheel Library (Python)
}
\seealso{
\code{\link[=libraries]{libraries()}}

Other Library Objects: 
\code{\link{lib_cran}()},
\code{\link{lib_egg}()},
\code{\link{lib_jar}()},
\code{\link{lib_maven}()},
\code{\link{lib_pypi}()},
\code{\link{libraries}()}
}
\concept{Library Objects}
