% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{libraries}
\alias{libraries}
\title{Libraries}
\usage{
libraries(...)
}
\arguments{
\item{...}{Accepts multiple instances of \code{\link[=lib_jar]{lib_jar()}}, \code{\link[=lib_cran]{lib_cran()}},
\code{\link[=lib_maven]{lib_maven()}}, \code{\link[=lib_pypi]{lib_pypi()}}, \code{\link[=lib_whl]{lib_whl()}}, \code{\link[=lib_egg]{lib_egg()}}.}
}
\description{
Libraries
}
\details{
Optional list of libraries to be installed on the cluster that executes the
task.
}
\seealso{
\code{\link[=job_task]{job_task()}}, \code{\link[=lib_jar]{lib_jar()}}, \code{\link[=lib_cran]{lib_cran()}}, \code{\link[=lib_maven]{lib_maven()}},
\code{\link[=lib_pypi]{lib_pypi()}}, \code{\link[=lib_whl]{lib_whl()}}, \code{\link[=lib_egg]{lib_egg()}}

Other Task Objects: 
\code{\link{condition_task}()},
\code{\link{email_notifications}()},
\code{\link{for_each_task}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{pipeline_task}()},
\code{\link{python_wheel_task}()},
\code{\link{run_job_task}()},
\code{\link{spark_jar_task}()},
\code{\link{spark_python_task}()},
\code{\link{spark_submit_task}()},
\code{\link{sql_file_task}()},
\code{\link{sql_query_task}()}

Other Library Objects: 
\code{\link{lib_cran}()},
\code{\link{lib_egg}()},
\code{\link{lib_jar}()},
\code{\link{lib_maven}()},
\code{\link{lib_pypi}()},
\code{\link{lib_whl}()}
}
\concept{Library Objects}
\concept{Task Objects}
