% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-helpers.R
\name{notebook_enable_htmlwidgets}
\alias{notebook_enable_htmlwidgets}
\title{Enable htmlwidgets in Databricks Notebook}
\usage{
notebook_enable_htmlwidgets(height = 450)
}
\arguments{
\item{height}{Measurement passed to height of htmlwidget. This overrides
existing values that may often be \code{NULL} to ensure the height is correctly
displayed within the iframe of notebook results cells (via \code{displayHTML()}).
Default is 450.}
}
\description{
Enable htmlwidgets in Databricks Notebook
}
\details{
Databricks notebooks by default don't currently support htmlwidgets.
This behaviour can be corrected by:
\itemize{
\item adjusting the print method in htmltools
\item installing pandoc
}

This is a invasive method to correct the behaviour as htmltools isn't
flexible to adjust via the \code{viewer} option directly.

It only runs within a Databricks notebook cell.

The height can be adjusted without running the function again by using the
\code{db_htmlwidget_height} option (e.g. \code{options(db_htmlwidget_height = 300)}).
}
\examples{
notebook_enable_htmlwidgets()
# set default height to 800px
notebook_enable_htmlwidgets(height = 800)
}
