% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook-helpers.R
\name{notebook_use_posit_repo}
\alias{notebook_use_posit_repo}
\title{Setup Databricks Notebook with Posit Package Manager}
\usage{
notebook_use_posit_repo()
}
\description{
Setup Databricks Notebook with Posit Package Manager
}
\details{
Databricks notebooks default repo for package installation is CRAN.
CRAN doesn't provide pre-compiled binaries for linux and this results in
packages taking longer than desired.

This function can be called within a Databricks notebook to easily switch to
Posit and retrieve pre-compiled binaries.

This function will behave correctly across different Databricks Runtimes,
even when the underlying linux version changes.
}
