% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{py_db_sql_connector}
\alias{py_db_sql_connector}
\title{Databricks SQL Connector (Python)}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
py_db_sql_connector
}
\description{
Databricks SQL Connector (Python)
}
\details{
TODO
}
\keyword{datasets}
