% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{read_env_var}
\alias{read_env_var}
\title{Reads Environment Variables}
\usage{
read_env_var(
  key = c("token", "host", "wsid", "client_id", "client_secret", "azure_client_id",
    "azure_client_secret", "azure_tenant_id", "auth_type"),
  profile = NULL,
  error = TRUE
)
}
\arguments{
\item{key}{The value to fetch from profile. One of \code{token}, \code{host}, \code{wsid},
\code{client_id}, \code{client_secret}, \code{azure_client_id}, \code{azure_client_secret},
\code{azure_tenant_id}, or \code{auth_type}}

\item{profile}{Character, the name of the profile to retrieve values}

\item{error}{Boolean, when key isn't found should error be raised}
}
\value{
named list of values associated with profile
}
\description{
Reads Environment Variables
}
\details{
Fetches relevant environment variables based on profile
}
\keyword{internal}
