% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-management.R
\name{remove_lib_path}
\alias{remove_lib_path}
\title{Remove Library Path}
\usage{
remove_lib_path(path, version = FALSE)
}
\arguments{
\item{path}{Directory to remove from \code{\link[=.libPaths]{.libPaths()}}.}

\item{version}{If \code{TRUE} will add the R version string to the end
of \code{path} before removal.}
}
\description{
Remove Library Path
}
\seealso{
\code{\link[=.libPaths]{.libPaths()}}, \code{\link[=remove_lib_path]{remove_lib_path()}}
}
