% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart_cluster.R
\name{restart_cluster}
\alias{restart_cluster}
\title{Restart a Databricks cluster}
\usage{
restart_cluster(cluster_id, workspace, token = NULL, verbose = T, ...)
}
\arguments{
\item{cluster_id}{A string containing the unique id for an online Databricks cluster}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{verbose}{If TRUE, will print the API response to the console.  Defaults to
FALSE.}

\item{...}{Additional options to be passed to \code{data.table::fread} which is used to
parse the API response.}
}
\value{
The API response
}
\description{
This function will restart an existing interactive Databricks cluster.
You can locate the cluster ID in the URL of the cluster configuration
page.  For example:
}
\details{
https://mycompany.cloud.databricks.com/#/setting/clusters/xxxx-xxxxx-xxxxxx/

Where xxxx-xxxxx-xxxxxx is the cluster ID.

The API endpoint for terminating a cluster is '2.0/clusters/restart'.
  For all details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
cluster_id <- "0818-155203-cheese22"

restart_cluster(workspace = workspace, cluster_id = cluster_id)
}
