% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_job.R
\name{run_job}
\alias{run_job}
\title{Run an existing job on Databricks.}
\usage{
run_job(job_id = NULL, name = NULL, workspace, token = NULL, verbose = T)
}
\arguments{
\item{job_id}{A number.  Generated by \code{\link{create_job}} or via
the Databricks web interface.}

\item{name}{Optional.  A string representing the name of the job.  If multiple
jobs share the same name, you'll need to rename the jobs or provide
the unique job ID.}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided, netrc
 will be used.}

\item{verbose}{If true, will pretty print the success or failure of the
request and add a run_id variable to the R environment.  Defaults to TRUE.}
}
\value{
The API response.
}
\description{
Takes a job_id and executes it on Databricks.  Must create a job first.
}
\details{
The API endpoint for running a job is '2.0/jobs/run-now'.  For all
details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
## Run a job
run_job(job_id = 206, workspace = 'https://eastus2.azuredatabricks.net', token = 'dapi1903210980d9a0ads0g9fas')

## Run a job by name
run_job(name = "My Unique Job Name", workspace = workspace, token = token)

}
