% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{run_job_task}
\alias{run_job_task}
\title{Run Job Task}
\usage{
run_job_task(job_id, job_parameters, full_refresh = FALSE)
}
\arguments{
\item{job_id}{ID of the job to trigger.}

\item{job_parameters}{Named list, job-level parameters used to trigger job.}

\item{full_refresh}{If the pipeline should perform a full refresh.}
}
\description{
Run Job Task
}
\seealso{
Other Task Objects: 
\code{\link{condition_task}()},
\code{\link{email_notifications}()},
\code{\link{for_each_task}()},
\code{\link{libraries}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{pipeline_task}()},
\code{\link{python_wheel_task}()},
\code{\link{spark_jar_task}()},
\code{\link{spark_python_task}()},
\code{\link{spark_submit_task}()},
\code{\link{sql_file_task}()},
\code{\link{sql_query_task}()}
}
\concept{Task Objects}
