% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs_list.R
\name{runs_list}
\alias{runs_list}
\title{List runs for a job from most recently to least.}
\usage{
runs_list(
  job_id = NULL,
  name = NULL,
  workspace,
  token = NULL,
  offset = 0,
  limit = 20,
  active_only = "false",
  completed_only = "false",
  verbose = TRUE
)
}
\arguments{
\item{job_id}{A number representing a valid run id generating by launching
a job on Databricks.}

\item{name}{Optional.  A string representing the name of the job.  If multiple
jobs share the same name, you'll need to rename the jobs or provide
the unique job ID.}

\item{workspace}{A string representing the web workspace of your Databricks
instance. E.g., "https://eastus2.azuredatabricks.net" or
"https://demo.cloud.databricks.com".}

\item{token}{A valid authentication token generated via User Settings in
Databricks or via the Databricks REST API 2.0.  If none is provided,
netrc will be used.}

\item{offset}{The number of the first run to return, relative to the most
recent run.  Defaults to 0 (the most recent run).}

\item{limit}{The number of runs to return.  Should be greater than 0 and
less than 150.  If 0 is used, the service will use the maximum limit.}

\item{active_only}{A string.  If 'true', only active runs will be
included in the results.  Cannot be true while completed_only is true.}

\item{completed_only}{A string.  If 'true', only completed runs will be
included in the results.  Cannot be true while active_only is true.}

\item{verbose}{If true, will pretty print the success or failure of the
request and add a run_id variable to the R environment.  Defaults to TRUE.}
}
\value{
A list with four elements:
\itemize{
    \item \emph{response} - The full API response, including status code.
    \item \emph{response_json} - The JSON result.
    \item \emph{response_list} - The JSON result transformed into a list.
    \item \emph{response_df} - The JSON result transformed into a dataframe.
}
}
\description{
Fetches all of the completed and active runs for a given job in the last 60 days.
Runs expire after 60 days, so see \code{runs_export} to save run results before expiry.
}
\details{
The API endpoint for running a job is '2.0/jobs/runs/list'.  For all
details on API calls please see the official documentation at
\url{https://docs.databricks.com/dev-tools/api/latest/}.
}
\examples{
## Get runs by Job ID
res <- runs_list(job_id = 206,
                      workspace = "https://eastus2.azuredatabricks.net",
                      token = "dapiafoi32984320890d9asaf1")

## Get runs by Job name
res <- runs_list(name = "My Other Unique Job Name", workspace = workspace, token = token)

## Get data in different formats
resjson <- res$response_json
reslist <- res$response_list
res_df <- res$response_df

}
