% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_library.R
\name{set_library}
\alias{set_library}
\alias{remove_library}
\title{Set the User Library Path}
\usage{
set_library(lib_path = "/dbfs/Rlib")

remove_library(lib_path)
}
\arguments{
\item{user_libpath}{A string representing the path to install packages in DBFS}

\item{user_lib_path}{A string representing the path to be removed from .libPaths()}
}
\value{
The user library path

The removed library path
}
\description{
Set the path for where R will install packages.
}
\details{
To persist packages on Databricks use a path that begins with `/dbfs/`.
This is designed to be used in conjunction with `bricksteR::curate()`.
For shared clusters where users prefer package isolation, each user should
create their own path for package installation.  If building a central
shared repo of packages, set the path to a common directory.
}
\examples{

path <- "/dbfs/rk/my_packages"

set_library(lib_path = path)

path <- "/dbfs/rk/my_packages"

remove_library(path)
}
