% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine-helpers.R
\name{setup_databricks_rmd}
\alias{setup_databricks_rmd}
\title{Setup Databricks RMarkdown}
\usage{
setup_databricks_rmd(cluster_id)
}
\arguments{
\item{cluster_id}{Databricks cluster ID}
}
\value{
Databricks execution context ID
}
\description{
Setup Databricks RMarkdown
}
\details{
This function should be run in an RMarkdown setup chunk to allow subsequent
chunks to use \verb{databricks_*} related knitr engines that allow remote
execution against a Databricks cluster.

This function deliberately does not have parameters to accept credentials
(host/token). These credentials should be managed elsewhere (e.g. \code{.Renviron})
and not stored in plain-text within the script or markdown.
}
