% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{show,DatabricksConnection-method}
\alias{show,DatabricksConnection-method}
\title{Show method for DatabricksConnection}
\usage{
\S4method{show}{DatabricksConnection}(object)
}
\arguments{
\item{object}{A DatabricksConnection object}
}
\description{
Show method for DatabricksConnection
}
