% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{show,DatabricksResult-method}
\alias{show,DatabricksResult-method}
\title{Show method for DatabricksResult}
\usage{
\S4method{show}{DatabricksResult}(object)
}
\arguments{
\item{object}{A DatabricksResult object}
}
\description{
Show method for DatabricksResult
}
