% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{spark_jar_task}
\alias{spark_jar_task}
\title{Spark Jar Task}
\usage{
spark_jar_task(main_class_name, parameters = list())
}
\arguments{
\item{main_class_name}{The full name of the class containing the main method
to be executed. This class must be contained in a JAR provided as a library.
The code must use \code{SparkContext.getOrCreate} to obtain a Spark context;
otherwise, runs of the job fail.}

\item{parameters}{Named list. Parameters passed to the main method. Use Task
parameter variables to set parameters containing information about job runs.}
}
\description{
Spark Jar Task
}
\seealso{
Other Task Objects: 
\code{\link{condition_task}()},
\code{\link{email_notifications}()},
\code{\link{for_each_task}()},
\code{\link{libraries}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{pipeline_task}()},
\code{\link{python_wheel_task}()},
\code{\link{run_job_task}()},
\code{\link{spark_python_task}()},
\code{\link{spark_submit_task}()},
\code{\link{sql_file_task}()},
\code{\link{sql_query_task}()}
}
\concept{Task Objects}
