% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{spark_python_task}
\alias{spark_python_task}
\title{Spark Python Task}
\usage{
spark_python_task(python_file, parameters = list())
}
\arguments{
\item{python_file}{The URI of the Python file to be executed. DBFS and S3
paths are supported.}

\item{parameters}{List. Command line parameters passed to the Python file.
Use Task parameter variables to set parameters containing information about
job runs.}
}
\description{
Spark Python Task
}
\seealso{
Other Task Objects: 
\code{\link{email_notifications}()},
\code{\link{libraries}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{pipeline_task}()},
\code{\link{python_wheel_task}()},
\code{\link{spark_jar_task}()},
\code{\link{spark_submit_task}()}
}
\concept{Task Objects}
