% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{sql_query_save.DatabricksConnection}
\alias{sql_query_save.DatabricksConnection}
\title{Create temporary views and tables in Databricks}
\usage{
\method{sql_query_save}{DatabricksConnection}(con, sql, name, temporary = TRUE, ...)
}
\arguments{
\item{con}{A DatabricksConnection object}

\item{sql}{SQL query to save as table/view}

\item{name}{Name for the temporary view or table}

\item{temporary}{Whether the object should be temporary (default: \code{TRUE})}

\item{...}{Additional arguments (ignored)}
}
\value{
The table/view name (invisibly)
}
\description{
Create temporary views and tables in Databricks
}
