% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{sql_query_task}
\alias{sql_query_task}
\title{SQL Query Task}
\usage{
sql_query_task(query_id, warehouse_id, parameters = NULL)
}
\arguments{
\item{query_id}{The canonical identifier of the SQL query.}

\item{warehouse_id}{The canonical identifier of the SQL warehouse.}

\item{parameters}{Named list of paramters to be used for each run of this job.}
}
\description{
SQL Query Task
}
\seealso{
Other Task Objects: 
\code{\link{condition_task}()},
\code{\link{email_notifications}()},
\code{\link{for_each_task}()},
\code{\link{libraries}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{pipeline_task}()},
\code{\link{python_wheel_task}()},
\code{\link{run_job_task}()},
\code{\link{spark_jar_task}()},
\code{\link{spark_python_task}()},
\code{\link{spark_submit_task}()},
\code{\link{sql_file_task}()}
}
\concept{Task Objects}
