% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{sql_translate_env.DatabricksConnection}
\alias{sql_translate_env.DatabricksConnection}
\title{SQL translation environment for Databricks SQL}
\usage{
sql_translate_env.DatabricksConnection(con)
}
\description{
SQL translation environment for Databricks SQL
}
\keyword{internal}
