% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{sql_translation.DatabricksConnection}
\alias{sql_translation.DatabricksConnection}
\title{SQL translation environment for Databricks SQL}
\usage{
\method{sql_translation}{DatabricksConnection}(con)
}
\description{
SQL translation environment for Databricks SQL
}
\keyword{internal}
