% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{warehouse_id_from_http_path}
\alias{warehouse_id_from_http_path}
\title{Extract warehouse ID from an http_path}
\usage{
warehouse_id_from_http_path(http_path)
}
\description{
Extract warehouse ID from an http_path
}
\keyword{internal}
