% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{auth_params}
\alias{auth_params}
\title{Authentication Parameters}
\arguments{
\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\description{
Shared host/token parameters inherited by brickster API wrappers.
}
\keyword{internal}
