% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{close_workspace}
\alias{close_workspace}
\title{Close Databricks Workspace Connection}
\usage{
close_workspace(host = db_host())
}
\arguments{
\item{host}{TODO}
}
\description{
Close Databricks Workspace Connection
}
\examples{
\dontrun{
close_workspace(host = db_host())
}
}
