% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbplyr.R
\name{databricks-dbplyr}
\alias{databricks-dbplyr}
\title{dbplyr Backend for Databricks SQL}
\description{
This file implements dbplyr backend support for Databricks SQL warehouses,
enabling dplyr syntax to be translated to Databricks SQL.
}
