% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-helpers.R
\name{db_current_cloud}
\alias{db_current_cloud}
\title{Detect Current Workspaces Cloud}
\usage{
db_current_cloud(host = db_host(), token = db_token(), perform_request = TRUE)
}
\arguments{
\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
String
}
\description{
Detect Current Workspaces Cloud
}
