% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{db_host}
\alias{db_host}
\title{Generate/Fetch Databricks Host}
\usage{
db_host(id = NULL, prefix = NULL, profile = getOption("db_profile", NULL))
}
\arguments{
\item{id}{The workspace string}

\item{prefix}{Workspace prefix}

\item{profile}{Profile to use when fetching from environment variable or
\code{.databricksfg} file}
}
\value{
workspace URL
}
\description{
If both \code{id} and \code{prefix} are \code{NULL} function will search for the
\code{DATABRICKS_HOST} environment variable.

When defining \code{id} and \code{prefix} you do not need to specify the whole URL.
E.g. \verb{https://<prefix>.<id>.cloud.databricks.com/} is the form to follow.
}
\seealso{
Other Databricks Authentication Helpers: 
\code{\link{db_read_netrc}()},
\code{\link{db_token}()},
\code{\link{db_wsid}()}
}
\concept{Databricks Authentication Helpers}
