% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakebase.R
\name{db_lakebase_get}
\alias{db_lakebase_get}
\title{Get Database Instance}
\usage{
db_lakebase_get(
  name,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{name}{Name of the database instance to retrieve.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
List
}
\description{
Get Database Instance
}
\seealso{
Other Database API: 
\code{\link{db_lakebase_creds_generate}()},
\code{\link{db_lakebase_get_by_uid}()},
\code{\link{db_lakebase_list}()}
}
\concept{Database API}
