% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{db_libs_cluster_status}
\alias{db_libs_cluster_status}
\title{Get Status of Libraries on Cluster}
\usage{
db_libs_cluster_status(
  cluster_id,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{cluster_id}{Unique identifier of a Databricks cluster.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Get Status of Libraries on Cluster
}
\seealso{
\code{\link[=wait_for_lib_installs]{wait_for_lib_installs()}}

Other Libraries API: 
\code{\link{db_libs_all_cluster_statuses}()},
\code{\link{db_libs_install}()},
\code{\link{db_libs_uninstall}()}
}
\concept{Libraries API}
