% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow-databricks.R
\name{db_mlflow_model_version_comment}
\alias{db_mlflow_model_version_comment}
\title{Make a Comment on a Model Version}
\usage{
db_mlflow_model_version_comment(
  name,
  version,
  comment,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{name}{Name of the model.}

\item{version}{Version of the model.}

\item{comment}{User-provided comment on the action.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\description{
Make a Comment on a Model Version
}
\seealso{
Other Model Registry API: 
\code{\link{db_mlflow_model_approve_transition_req}()},
\code{\link{db_mlflow_model_delete_transition_req}()},
\code{\link{db_mlflow_model_open_transition_reqs}()},
\code{\link{db_mlflow_model_reject_transition_req}()},
\code{\link{db_mlflow_model_transition_req}()},
\code{\link{db_mlflow_model_transition_stage}()},
\code{\link{db_mlflow_model_version_comment_delete}()},
\code{\link{db_mlflow_model_version_comment_edit}()},
\code{\link{db_mlflow_registered_model_details}()}
}
\concept{Model Registry API}
