% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlflow-databricks.R
\name{db_mlflow_model_version_comment_edit}
\alias{db_mlflow_model_version_comment_edit}
\title{Edit a Comment on a Model Version}
\usage{
db_mlflow_model_version_comment_edit(
  id,
  comment,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{id}{Unique identifier of an activity.}

\item{comment}{User-provided comment on the action.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Edit a Comment on a Model Version
}
\seealso{
Other Model Registry API: 
\code{\link{db_mlflow_model_approve_transition_req}()},
\code{\link{db_mlflow_model_delete_transition_req}()},
\code{\link{db_mlflow_model_open_transition_reqs}()},
\code{\link{db_mlflow_model_reject_transition_req}()},
\code{\link{db_mlflow_model_transition_req}()},
\code{\link{db_mlflow_model_transition_stage}()},
\code{\link{db_mlflow_model_version_comment_delete}()},
\code{\link{db_mlflow_model_version_comment}()},
\code{\link{db_mlflow_registered_model_details}()}
}
\concept{Model Registry API}
