% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{db_oauth_client}
\alias{db_oauth_client}
\title{Create OAuth 2.0 Client}
\usage{
db_oauth_client(host = db_host())
}
\arguments{
\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}
}
\value{
List that contains httr2_oauth_client and relevant auth url
}
\description{
Create OAuth 2.0 Client
}
\details{
Creates an OAuth 2.0 Client, support for U2M flows only.
May later be extended for account U2M and all M2M flows.
}
\keyword{internal}
