% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-helpers.R
\name{db_perform_request}
\alias{db_perform_request}
\title{Perform Databricks API Request}
\usage{
db_perform_request(req, ...)
}
\arguments{
\item{req}{\code{{httr2}} request.}

\item{...}{Parameters passed to \code{\link[httr2:resp_body_raw]{httr2::resp_body_json()}}}
}
\description{
Perform Databricks API Request
}
\seealso{
Other Request Helpers: 
\code{\link{db_perform_response}()},
\code{\link{db_req_error_body}()},
\code{\link{db_request}()},
\code{\link{db_request_json}()}
}
\concept{Request Helpers}
