% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{db_query_get}
\alias{db_query_get}
\title{Get a SQL Query}
\usage{
db_query_get(id, host = db_host(), token = db_token(), perform_request = TRUE)
}
\arguments{
\item{id}{String, ID for the query.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Returns the repo with the given repo ID.
}
\seealso{
Other SQL Queries API: 
\code{\link{db_query_create}()},
\code{\link{db_query_delete}()},
\code{\link{db_query_list}()},
\code{\link{db_query_update}()}
}
\concept{SQL Queries API}
