% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{db_read_netrc}
\alias{db_read_netrc}
\title{Read .netrc File}
\usage{
db_read_netrc(path = "~/.netrc")
}
\arguments{
\item{path}{path of \code{.netrc} file, default is \verb{~/.netrc}.}
}
\value{
named list of \code{.netrc} entries
}
\description{
Read .netrc File
}
\seealso{
Other Databricks Authentication Helpers: 
\code{\link{db_host}()},
\code{\link{db_token}()}
}
\concept{Databricks Authentication Helpers}
