% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-repl.R
\name{db_repl}
\alias{db_repl}
\title{Remote REPL to Databricks Cluster}
\usage{
db_repl(
  cluster_id,
  language = c("r", "py", "scala", "sql", "sh"),
  host = db_host(),
  token = db_token()
)
}
\arguments{
\item{cluster_id}{Cluster Id to create REPL context against.}

\item{language}{for REPL ('r', 'py', 'scala', 'sql', 'sh') are
supported.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\description{
Remote REPL to Databricks Cluster
}
\details{
This function doesn't accept \code{token} and \code{host} parameters,
credentials must be established as per documentation best practice
(e.g. \code{.Renviron}).

\code{db_repl()} will take over the existing console and allow execution of
commands against a Databricks cluster. For RStudio users there are Addins
which can be bound to keyboard shortcuts to improve usability.
}
