% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{db_repo_get}
\alias{db_repo_get}
\title{Get Repo}
\usage{
db_repo_get(
  repo_id,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{repo_id}{The ID for the corresponding repo to access.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
Returns the repo with the given repo ID.
}
\seealso{
Other Repos API: 
\code{\link{db_repo_create}()},
\code{\link{db_repo_delete}()},
\code{\link{db_repo_get_all}()},
\code{\link{db_repo_update}()}
}
\concept{Repos API}
