% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-helpers.R
\name{db_request}
\alias{db_request}
\title{Databricks Request Helper}
\usage{
db_request(endpoint, method, version = NULL, body = NULL, host, token, ...)
}
\arguments{
\item{endpoint}{Databricks REST API Endpoint}

\item{method}{Passed to \code{\link[httr2:req_method]{httr2::req_method()}}}

\item{version}{String, API version of endpoint. E.g. \code{2.0}.}

\item{body}{Named list, passed to \code{\link[httr2:req_body]{httr2::req_body_json()}}.}

\item{host}{Databricks host, defaults to \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks token, defaults to \code{\link[=db_token]{db_token()}}.}

\item{...}{Parameters passed on to \code{\link[httr2:req_body]{httr2::req_body_json()}} when \code{body} is not \code{NULL}.}
}
\value{
request
}
\description{
Databricks Request Helper
}
\details{
If you need to pass advanced options to the request, you can use the
option \code{"brickster.request.options"}. For instance, to force HTTP/1.1 you can
use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options("brickster.request.options" = list(
  http_version = curl::curl_symbols("CURL_HTTP_VERSION_1_1")$value)
)
}\if{html}{\out{</div>}}

These options will be passed to \code{\link[httr2:req_options]{httr2::req_options()}}.
}
\seealso{
Other Request Helpers: 
\code{\link{db_perform_request}()},
\code{\link{db_req_error_body}()},
\code{\link{db_request_json}()}
}
\concept{Request Helpers}
