% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-helpers.R
\name{db_request_json}
\alias{db_request_json}
\title{Generate Request JSON}
\usage{
db_request_json(req)
}
\arguments{
\item{req}{a httr2 request, ideally from \code{\link[=db_request]{db_request()}}.}
}
\value{
JSON string
}
\description{
Generate Request JSON
}
\seealso{
Other Request Helpers: 
\code{\link{db_perform_request}()},
\code{\link{db_req_error_body}()},
\code{\link{db_request}()}
}
\concept{Request Helpers}
