% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-connector.R
\name{db_sql_client}
\alias{db_sql_client}
\title{Create Databricks SQL Connector Client}
\usage{
db_sql_client(
  id,
  catalog = NULL,
  schema = NULL,
  compute_type = c("warehouse", "cluster"),
  use_cloud_fetch = FALSE,
  session_configuration = list(),
  host = db_host(),
  token = db_token(),
  workspace_id = db_wsid(),
  ...
)
}
\arguments{
\item{id}{TODO}

\item{catalog}{TODO}

\item{schema}{TODO}

\item{compute_type}{TODO}

\item{use_cloud_fetch}{TODO}

\item{session_configuration}{TODO}

\item{host}{TODO}

\item{token}{TODO}

\item{workspace_id}{TODO}
}
\value{
TODO
}
\description{
Create Databricks SQL Connector Client
}
\details{
TODO
}
\examples{
TODO
}
