% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-query-execution.R
\name{db_sql_exec_poll_for_success}
\alias{db_sql_exec_poll_for_success}
\title{Poll a Query Until Successful}
\usage{
db_sql_exec_poll_for_success(statement_id, interval = 1)
}
\arguments{
\item{statement_id}{String, query execution \code{statement_id}}

\item{interval}{Number of seconds between status checks.}
}
\description{
Poll a Query Until Successful
}
