% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-query-execution.R
\name{db_sql_query}
\alias{db_sql_query}
\title{Execute SQL Query}
\usage{
db_sql_query(
  statement,
  warehouse_id,
  catalog = NULL,
  schema = NULL,
  parameters = NULL,
  row_limit = NULL,
  byte_limit = NULL,
  disposition = c("INLINE", "EXTERNAL_LINKS"),
  format = c("JSON_ARRAY", "ARROW_STREAM", "CSV"),
  wait_timeout = "10s",
  on_wait_timeout = c("CONTINUE", "CANCEL"),
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{statement}{TODO}

\item{warehouse_id}{TODO}

\item{catalog}{TODO}

\item{schema}{TODO}

\item{parameters}{TODO}

\item{row_limit}{TODO}

\item{byte_limit}{TODO}

\item{disposition}{TODO}

\item{format}{TODO}

\item{wait_timeout}{TODO}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\description{
Execute SQL Query
}
\details{
TODO
}
\seealso{
Other SQL Execution APIs: 
\code{\link{db_sql_cancel}()},
\code{\link{db_sql_result}()},
\code{\link{db_sql_status}()}
}
\concept{SQL Execution APIs}
