% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouses.R
\name{db_sql_warehouse_list}
\alias{db_sql_warehouse_list}
\title{List Warehouses}
\usage{
db_sql_warehouse_list(
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
List Warehouses
}
\seealso{
Other Warehouse API: 
\code{\link{db_sql_global_warehouse_edit}()},
\code{\link{db_sql_global_warehouse_get}()},
\code{\link{db_sql_warehouse_create}()},
\code{\link{db_sql_warehouse_delete}()},
\code{\link{db_sql_warehouse_edit}()},
\code{\link{db_sql_warehouse_get}()},
\code{\link{db_sql_warehouse_start}()},
\code{\link{db_sql_warehouse_stop}()},
\code{\link{get_and_start_warehouse}()}
}
\concept{Warehouse API}
