% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{db_token}
\alias{db_token}
\title{Fetch Databricks Token}
\usage{
db_token(profile = getOption("db_profile"))
}
\arguments{
\item{profile}{Profile to use when fetching from environment variable
(e.g. \code{.Renviron}) or \code{.databricksfg} file}
}
\value{
databricks token
}
\description{
Token must be specified as an environment variable \code{DATABRICKS_TOKEN}.
If \code{DATABRICKS_TOKEN} is missing will default to using OAuth U2M flow.

Refer to \href{https://docs.databricks.com/dev-tools/api/latest/authentication.html}{api authentication docs}
}
\details{
The behaviour is subject to change depending if \code{db_profile} and
\code{use_databrickscfg} options are set.
\itemize{
\item \code{use_databrickscfg}: Boolean (default: \code{FALSE}), determines if credentials
are fetched from profile of \code{.databrickscfg} or \code{.Renviron}
\item \code{db_profile}: String (default: \code{NULL}), determines profile used.
}

See vignette on authentication for more details.
}
\seealso{
Other Databricks Authentication Helpers: 
\code{\link{db_host}()},
\code{\link{db_read_netrc}()},
\code{\link{db_wsid}()}
}
\concept{Databricks Authentication Helpers}
