% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{db_token}
\alias{db_token}
\title{Fetch Databricks Token}
\usage{
db_token(profile = getOption("db_profile"))
}
\arguments{
\item{profile}{Profile to use when fetching from environment variable or
\code{.databricksfg} file}
}
\value{
databricks token
}
\description{
Token must be specified as an environment variable \code{DATABRICKS_TOKEN}.

Refer to \href{https://docs.databricks.com/dev-tools/api/latest/authentication.html}{api authentication docs}
}
\seealso{
Other Databricks Authentication Helpers: 
\code{\link{db_host}()},
\code{\link{db_read_netrc}()},
\code{\link{db_wsid}()}
}
\concept{Databricks Authentication Helpers}
