% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{db_token}
\alias{db_token}
\title{Fetch Databricks Token}
\usage{
db_token(profile = default_config_profile())
}
\arguments{
\item{profile}{Profile to use when fetching from environment variable
(e.g. \code{.Renviron}) or \code{.databricksfg} file}
}
\value{
databricks token
}
\description{
The function will check for a token in the \code{DATABRICKS_HOST} environment variable.
\code{.databrickscfg} will be searched if \code{db_profile} and \code{use_databrickscfg} are set or
if Posit Workbench managed OAuth credentials are detected. If none of the above are
found then \code{db_token()} returns \code{NULL}.

Refer to \href{https://docs.databricks.com/aws/en/dev-tools/auth}{api authentication docs}
}
\details{
The behaviour is subject to change depending if \code{db_profile} and
\code{use_databrickscfg} options are set.
\itemize{
\item \code{use_databrickscfg}: Boolean (default: \code{FALSE}), determines if credentials
are fetched from profile of \code{.databrickscfg} or \code{.Renviron}
\item \code{db_profile}: String (default: \code{NULL}), determines profile used.
\code{.databrickscfg} will automatically be used when Posit Workbench managed OAuth credentials are detected.
}

See vignette on authentication for more details.
}
\seealso{
Other Databricks Authentication Helpers: 
\code{\link{db_host}()},
\code{\link{db_read_netrc}()},
\code{\link{db_token_provider}},
\code{\link{db_wsid}()}
}
\concept{Databricks Authentication Helpers}
