% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-schemas.R
\name{db_uc_schemas_get}
\alias{db_uc_schemas_get}
\title{Get Schema (Unity Catalog)}
\usage{
db_uc_schemas_get(
  catalog,
  schema,
  include_browse = TRUE,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{catalog}{Parent catalog for schema of interest.}

\item{schema}{Schema of interest.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
List
}
\description{
Get Schema (Unity Catalog)
}
