% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc-tables.R
\name{db_uc_tables_exists}
\alias{db_uc_tables_exists}
\title{Check Table Exists (Unity Catalog)}
\usage{
db_uc_tables_exists(
  catalog,
  schema,
  table,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{catalog}{Parent catalog of table.}

\item{schema}{Parent schema of table.}

\item{table}{Table name.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
List with fields \code{table_exists} and \code{supports_foreign_metadata_update}
}
\description{
Check Table Exists (Unity Catalog)
}
\seealso{
Other Unity Catalog Table Management: 
\code{\link{db_uc_tables_delete}()},
\code{\link{db_uc_tables_get}()},
\code{\link{db_uc_tables_list}()},
\code{\link{db_uc_tables_summaries}()}
}
\concept{Unity Catalog Table Management}
