% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-fs.R
\name{db_volume_dir_delete}
\alias{db_volume_dir_delete}
\title{Volume FileSystem Delete Directory}
\usage{
db_volume_dir_delete(
  path,
  recursive = FALSE,
  verbose = FALSE,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{path}{Absolute path of the file in the Files API, omitting the initial
slash.}

\item{recursive}{If \code{TRUE}, recursively delete directory contents (default: \code{FALSE})}

\item{verbose}{If \code{TRUE}, announce each file/directory deletion (default: \code{FALSE})}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\value{
If \code{perform_request = TRUE}, returns a logical success flag.
If \code{FALSE}, returns an \code{httr2_request} to delete only the directory itself.
Directory contents are not listed or deleted, even when \code{recursive = TRUE}.
}
\description{
Volume FileSystem Delete Directory
}
\seealso{
Other Volumes FileSystem API: 
\code{\link{db_volume_delete}()},
\code{\link{db_volume_dir_create}()},
\code{\link{db_volume_dir_exists}()},
\code{\link{db_volume_download_dir}()},
\code{\link{db_volume_file_exists}()},
\code{\link{db_volume_list}()},
\code{\link{db_volume_read}()},
\code{\link{db_volume_upload_dir}()},
\code{\link{db_volume_write}()}
}
\concept{Volumes FileSystem API}
