% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-fs.R
\name{db_volume_download_dir}
\alias{db_volume_download_dir}
\title{Download Directory from Volume in Parallel}
\usage{
db_volume_download_dir(
  volume_dir,
  local_dir,
  overwrite = TRUE,
  recursive = TRUE,
  host = db_host(),
  token = db_token()
)
}
\arguments{
\item{volume_dir}{Volume directory path (must start with /Volumes/)}

\item{local_dir}{Path to local directory where files will be downloaded}

\item{overwrite}{Flag to overwrite existing local files (default: \code{TRUE})}

\item{recursive}{If \code{TRUE}, recursively include subdirectories (default: \code{TRUE}).
If \code{FALSE}, only top-level files are transferred.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\value{
TRUE if all downloads successful
}
\description{
Download files from a volume directory to a local directory using parallel requests.
}
\seealso{
Other Volumes FileSystem API: 
\code{\link{db_volume_delete}()},
\code{\link{db_volume_dir_create}()},
\code{\link{db_volume_dir_delete}()},
\code{\link{db_volume_dir_exists}()},
\code{\link{db_volume_file_exists}()},
\code{\link{db_volume_list}()},
\code{\link{db_volume_read}()},
\code{\link{db_volume_upload_dir}()},
\code{\link{db_volume_write}()}
}
\concept{Volumes FileSystem API}
