% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume-fs.R
\name{db_volume_write}
\alias{db_volume_write}
\title{Volume FileSystem Write}
\usage{
db_volume_write(
  path,
  file = NULL,
  contents = NULL,
  overwrite = FALSE,
  host = db_host(),
  token = db_token()
)
}
\arguments{
\item{path}{Absolute path of the file in the Files API, omitting the initial
slash.}

\item{file}{Path to a file on local system, takes precedent over \code{path}.}

\item{contents}{String that is base64 encoded.}

\item{overwrite}{Flag (Default: \code{FALSE}) that specifies whether to overwrite
existing files.}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}
}
\description{
Upload a file to volume filesystem.
}
\details{
Either \code{contents} or \code{file} must be specified. \code{file} takes precedent over
\code{contents} if both are specified.

Uploads a file of up to 2 GiB.
}
\seealso{
Other Volumes FileSystem API: 
\code{\link{db_volume_delete}()},
\code{\link{db_volume_file_exists}()},
\code{\link{db_volume_read}()}
}
\concept{Volumes FileSystem API}
