% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-search.R
\name{db_vs_indexes_list}
\alias{db_vs_indexes_list}
\title{List Vector Search Indexes}
\usage{
db_vs_indexes_list(
  endpoint,
  page_token = NULL,
  host = db_host(),
  token = db_token(),
  perform_request = TRUE
)
}
\arguments{
\item{endpoint}{Name of vector search endpoint}

\item{page_token}{\code{page_token} returned from prior query}

\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{perform_request}{If \code{TRUE} (default) the request is performed, if
\code{FALSE} the httr2 request is returned \emph{without} being performed.}
}
\description{
List Vector Search Indexes
}
\seealso{
Other Vector Search API: 
\code{\link{db_vs_endpoints_create}()},
\code{\link{db_vs_endpoints_delete}()},
\code{\link{db_vs_endpoints_get}()},
\code{\link{db_vs_endpoints_list}()},
\code{\link{db_vs_indexes_create}()},
\code{\link{db_vs_indexes_delete}()},
\code{\link{db_vs_indexes_delete_data}()},
\code{\link{db_vs_indexes_get}()},
\code{\link{db_vs_indexes_query}()},
\code{\link{db_vs_indexes_query_next_page}()},
\code{\link{db_vs_indexes_scan}()},
\code{\link{db_vs_indexes_sync}()},
\code{\link{db_vs_indexes_upsert_data}()},
\code{\link{delta_sync_index_spec}()},
\code{\link{direct_access_index_spec}()},
\code{\link{embedding_source_column}()},
\code{\link{embedding_vector_column}()}
}
\concept{Vector Search API}
