% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{db_wsid}
\alias{db_wsid}
\title{Fetch Databricks Workspace ID}
\usage{
db_wsid(profile = getOption("db_profile"))
}
\arguments{
\item{profile}{Profile to use when fetching from environment variable or
\code{.databricksfg} file}
}
\value{
databricks workspace ID
}
\description{
Workspace ID, optionally specificied to make connections pane more powerful.
Specified as an environment variable \code{DATABRICKS_WSID}.

Refer to \href{https://docs.databricks.com/dev-tools/api/latest/authentication.html}{api authentication docs}
}
\seealso{
Other Databricks Authentication Helpers: 
\code{\link{db_host}()},
\code{\link{db_read_netrc}()},
\code{\link{db_token}()}
}
\concept{Databricks Authentication Helpers}
