% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbClearResult,DatabricksResult-method}
\alias{dbClearResult,DatabricksResult-method}
\title{Clear result set}
\usage{
\S4method{dbClearResult}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
TRUE (invisibly)
}
\description{
Clear result set
}
