% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbDisconnect,DatabricksConnection-method}
\alias{dbDisconnect,DatabricksConnection-method}
\title{Disconnect from Databricks}
\usage{
\S4method{dbDisconnect}{DatabricksConnection}(conn, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{...}{Additional arguments (ignored)}
}
\value{
TRUE (invisibly)
}
\description{
Disconnect from Databricks
}
