% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbExecute,DatabricksConnection,character-method}
\alias{dbExecute,DatabricksConnection,character-method}
\title{Execute statement on Databricks (not supported)}
\usage{
\S4method{dbExecute}{DatabricksConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{statement}{SQL statement}

\item{...}{Additional arguments (ignored)}
}
\value{
Always throws an error (read-only connection)
}
\description{
Execute statement on Databricks (not supported)
}
