% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbExecute,DatabricksConnection,character-method}
\alias{dbExecute,DatabricksConnection,character-method}
\title{Execute statement on Databricks}
\usage{
\S4method{dbExecute}{DatabricksConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{statement}{SQL statement}

\item{...}{Additional arguments (ignored)}
}
\value{
Number of rows in result set (from metadata, without loading data)
}
\description{
Execute statement on Databricks
}
