% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbFetch,DatabricksResult-method}
\alias{dbFetch,DatabricksResult-method}
\title{Fetch results from Databricks query}
\usage{
\S4method{dbFetch}{DatabricksResult}(res, n = -1, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{n}{Maximum number of rows to fetch (-1 for all rows)}

\item{...}{Additional arguments (ignored)}
}
\value{
A data.frame with query results
}
\description{
Fetch results from Databricks query
}
