% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbGetQuery,DatabricksConnection,character-method}
\alias{dbGetQuery,DatabricksConnection,character-method}
\title{Execute SQL query and return results}
\usage{
\S4method{dbGetQuery}{DatabricksConnection,character}(conn, statement, disposition = "EXTERNAL_LINKS", ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{statement}{SQL statement to execute}

\item{disposition}{Query disposition mode: "EXTERNAL_LINKS" (default) for large results,
"INLINE" for small metadata queries (automatically chooses appropriate format)}

\item{...}{Additional arguments passed to underlying query execution}
}
\value{
A data.frame with query results
}
\description{
Execute SQL query and return results
}
