% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbGetRowsAffected,DatabricksResult-method}
\alias{dbGetRowsAffected,DatabricksResult-method}
\title{Get number of rows affected (not applicable for SELECT)}
\usage{
\S4method{dbGetRowsAffected}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
-1 (not applicable for SELECT queries)
}
\description{
Get number of rows affected (not applicable for SELECT)
}
