% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbHasCompleted,DatabricksResult-method}
\alias{dbHasCompleted,DatabricksResult-method}
\title{Check whether all query rows have been fetched}
\usage{
\S4method{dbHasCompleted}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{TRUE} when no rows remain to be fetched, \code{FALSE} otherwise. Server
execution completing does not complete a result that still has unread rows.
Results from \code{\link[=dbSendStatement]{dbSendStatement()}} always return \code{TRUE}.
}
\description{
Check whether all query rows have been fetched
}
