% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbHasCompleted,DatabricksResult-method}
\alias{dbHasCompleted,DatabricksResult-method}
\title{Check if query has completed}
\usage{
\S4method{dbHasCompleted}{DatabricksResult}(res, ...)
}
\arguments{
\item{res}{A DatabricksResult object}

\item{...}{Additional arguments (ignored)}
}
\value{
TRUE if query is complete, FALSE otherwise
}
\description{
Check if query has completed
}
