% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks-dbi.R
\name{dbSendStatement,DatabricksConnection,character-method}
\alias{dbSendStatement,DatabricksConnection,character-method}
\title{Send statement to Databricks (not supported)}
\usage{
\S4method{dbSendStatement}{DatabricksConnection,character}(conn, statement, ...)
}
\arguments{
\item{conn}{A DatabricksConnection object}

\item{statement}{SQL statement}

\item{...}{Additional arguments (ignored)}
}
\value{
Always throws an error (read-only connection)
}
\description{
Send statement to Databricks (not supported)
}
