% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-connector.R
\name{install_db_sql_connector}
\alias{install_db_sql_connector}
\title{Install Databricks SQL Connector (Python)}
\usage{
install_db_sql_connector(envname = "r-brickster", method = "auto", ...)
}
\arguments{
\item{envname}{TODO}

\item{method}{TODO}

\item{...}{TODO}
}
\description{
Install Databricks SQL Connector (Python)
}
\details{
TODO
}
\examples{
TODO
}
