% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{job_tasks}
\alias{job_tasks}
\title{Job Tasks}
\usage{
job_tasks(...)
}
\arguments{
\item{...}{Multiple Instance of tasks \code{\link[=job_task]{job_task()}}.}
}
\description{
Job Tasks
}
\seealso{
\code{\link[=db_jobs_create]{db_jobs_create()}}, \code{\link[=db_jobs_reset]{db_jobs_reset()}}, \code{\link[=db_jobs_update]{db_jobs_update()}}
}
