% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{open_workspace}
\alias{open_workspace}
\title{Connect to Databricks Workspace}
\usage{
open_workspace(host = db_host(), token = db_token(), name = NULL)
}
\arguments{
\item{host}{TODO}

\item{token}{TODO}

\item{name}{TODO}
}
\description{
Connect to Databricks Workspace
}
\examples{
\dontrun{
open_workspace(host = db_host(), token = db_token, name = "MyWorkspace")
}
}
