% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{open_workspace}
\alias{open_workspace}
\title{Connect to Databricks Workspace}
\usage{
open_workspace(host = db_host(), token = db_token(), name = db_wsid())
}
\arguments{
\item{host}{Databricks workspace URL, defaults to calling \code{\link[=db_host]{db_host()}}.}

\item{token}{Databricks workspace token, defaults to calling \code{\link[=db_token]{db_token()}}.}

\item{name}{Desired name to assign the connection}
}
\description{
Connect to Databricks Workspace
}
\examples{
\dontrun{
open_workspace(host = db_host(), token = db_token, name = "MyWorkspace")
}
}
