% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{pipeline_task}
\alias{pipeline_task}
\title{Pipeline Task}
\usage{
pipeline_task(pipeline_id)
}
\arguments{
\item{pipeline_id}{The full name of the pipeline task to execute.}
}
\description{
Pipeline Task
}
\seealso{
Other Task Objects: 
\code{\link{email_notifications}()},
\code{\link{libraries}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{python_wheel_task}()},
\code{\link{spark_jar_task}()},
\code{\link{spark_python_task}()},
\code{\link{spark_submit_task}()}
}
\concept{Task Objects}
