% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{py_db_sql_connector}
\alias{py_db_sql_connector}
\title{Databricks SQL Connector (Python)}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
py_db_sql_connector
}
\description{
Access the Databricks SQL connector from Python via
\code{{reticulate}}.
}
\details{
This requires that the connector has been installed via
\code{\link[=install_db_sql_connector]{install_db_sql_connector()}}.

For more documentation of the methods, refer to the
\href{https://github.com/databricks/databricks-sql-python}{python documentation}.
}
\keyword{datasets}
