% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-auth.R
\name{read_databrickscfg}
\alias{read_databrickscfg}
\title{Reads Databricks CLI Config}
\usage{
read_databrickscfg(key = c("token", "host", "wsid"), profile = NULL)
}
\arguments{
\item{key}{The value to fetch from profile. One of \code{token}, \code{host}, or \code{wsid}}

\item{profile}{Character, the name of the profile to retrieve values}
}
\value{
named list of values associated with profile
}
\description{
Reads Databricks CLI Config
}
\details{
Reads \code{.databrickscfg} file and retrieves the values associated to
a given profile. Brickster searches for this file in the home directory.
}
