% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{spark_submit_task}
\alias{spark_submit_task}
\title{Spark Submit Task}
\usage{
spark_submit_task(parameters)
}
\arguments{
\item{parameters}{List. Command-line parameters passed to spark submit. Use
Task parameter variables to set parameters containing information about job runs.}
}
\description{
Spark Submit Task
}
\seealso{
Other Task Objects: 
\code{\link{email_notifications}()},
\code{\link{libraries}()},
\code{\link{new_cluster}()},
\code{\link{notebook_task}()},
\code{\link{pipeline_task}()},
\code{\link{python_wheel_task}()},
\code{\link{spark_jar_task}()},
\code{\link{spark_python_task}()}
}
\concept{Task Objects}
